Revert "Extend the schizo framework"
