+++ orte/mca/schizo/base/schizo_base_frame.c
@@ -37,7 +37,6 @@
  */
 orte_schizo_base_t orte_schizo_base = {{{0}}};
 orte_schizo_base_module_t orte_schizo = {
-    .define_cli = orte_schizo_base_define_cli,
     .parse_cli = orte_schizo_base_parse_cli,
     .parse_env = orte_schizo_base_parse_env,
     .setup_app = orte_schizo_base_setup_app,
@@ -47,28 +46,10 @@ orte_schizo_base_module_t orte_schizo = {
     .finalize = orte_schizo_base_finalize
 };
 
-static char *personalities = NULL;
-
-static int orte_schizo_base_register(mca_base_register_flag_t flags)
-{
-    /* pickup any defined personalities */
-    personalities = NULL;
-    mca_base_var_register("orte", "schizo", "base", "personalities",
-                          "Comma-separated list of personalities",
-                          MCA_BASE_VAR_TYPE_STRING, NULL, 0, 0,
-                          OPAL_INFO_LVL_9,
-                          MCA_BASE_VAR_SCOPE_READONLY,
-                          &personalities);
-    return ORTE_SUCCESS;
-}
-
 static int orte_schizo_base_close(void)
 {
     /* cleanup globals */
     OPAL_LIST_DESTRUCT(&orte_schizo_base.active_modules);
-    if (NULL != orte_schizo_base.personalities) {
-        opal_argv_free(orte_schizo_base.personalities);
-    }
 
     return mca_base_framework_components_close(&orte_schizo_base_framework, NULL);
 }
@@ -83,10 +64,6 @@ static int orte_schizo_base_open(mca_base_open_flag_t flags)
 
     /* init the globals */
     OBJ_CONSTRUCT(&orte_schizo_base.active_modules, opal_list_t);
-    orte_schizo_base.personalities = NULL;
-    if (NULL != personalities) {
-        orte_schizo_base.personalities = opal_argv_split(personalities, ',');
-    }
 
     /* Open up all available components */
     rc = mca_base_framework_components_open(&orte_schizo_base_framework, flags);
@@ -96,8 +73,7 @@ static int orte_schizo_base_open(mca_base_open_flag_t flags)
 }
 
 MCA_BASE_FRAMEWORK_DECLARE(orte, schizo, "ORTE Schizo Subsystem",
-                           orte_schizo_base_register,
-                           orte_schizo_base_open, orte_schizo_base_close,
+                           NULL, orte_schizo_base_open, orte_schizo_base_close,
                            mca_schizo_base_static_components, 0);
 
 OBJ_CLASS_INSTANCE(orte_schizo_base_active_module_t,
