+++ orte/mca/schizo/ompi/schizo_ompi.c
@@ -13,7 +13,7 @@
  *                         All rights reserved.
  * Copyright (c) 2009-2015 Cisco Systems, Inc.  All rights reserved.
  * Copyright (c) 2011      Oak Ridge National Labs.  All rights reserved.
- * Copyright (c) 2013-2016 Intel, Inc.  All rights reserved.
+ * Copyright (c) 2013-2015 Intel, Inc.  All rights reserved.
  * Copyright (c) 2015      Research Organization for Information Science
  *                         and Technology (RIST). All rights reserved.
  * $COPYRIGHT$
@@ -42,17 +42,17 @@
 #include "orte/mca/errmgr/errmgr.h"
 #include "orte/mca/ess/base/base.h"
 #include "orte/mca/rmaps/rmaps_types.h"
-#include "orte/orted/orted_submit.h"
 #include "orte/util/name_fns.h"
 #include "orte/util/session_dir.h"
 #include "orte/util/show_help.h"
 #include "orte/runtime/orte_globals.h"
 
-#include "orte/mca/schizo/base/base.h"
+#include "orte/mca/schizo/schizo.h"
 
-static int define_cli(opal_cmd_line_t *cli);
-static int parse_cli(int argc, int start, char **argv);
-static int parse_env(char *path,
+static int parse_cli(char **personality,
+                     int argc, int start, char **argv);
+static int parse_env(char **personality,
+                     char *path,
                      opal_cmd_line_t *cmd_line,
                      char **srcenv,
                      char ***dstenv);
@@ -63,426 +63,14 @@ static int setup_child(orte_job_t *jobdat,
                        orte_app_context_t *app);
 
 orte_schizo_base_module_t orte_schizo_ompi_module = {
-    .define_cli = define_cli,
     .parse_cli = parse_cli,
     .parse_env = parse_env,
     .setup_fork = setup_fork,
     .setup_child = setup_child
 };
 
-
-static opal_cmd_line_init_t cmd_line_init[] = {
-    /* Various "obvious" options */
-    { NULL, 'h', NULL, "help", 0,
-      &orte_cmd_line.help, OPAL_CMD_LINE_TYPE_BOOL,
-      "This help message" },
-    { NULL, 'V', NULL, "version", 0,
-      &orte_cmd_line.version, OPAL_CMD_LINE_TYPE_BOOL,
-      "Print version and exit" },
-    { NULL, 'v', NULL, "verbose", 0,
-      &orte_cmd_line.verbose, OPAL_CMD_LINE_TYPE_BOOL,
-      "Be verbose" },
-    { "orte_execute_quiet", 'q', NULL, "quiet", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Suppress helpful messages" },
-    { NULL, '\0', "report-pid", "report-pid", 1,
-      &orte_cmd_line.report_pid, OPAL_CMD_LINE_TYPE_STRING,
-      "Printout pid on stdout [-], stderr [+], or a file [anything else]" },
-    { NULL, '\0', "report-uri", "report-uri", 1,
-      &orte_cmd_line.report_uri, OPAL_CMD_LINE_TYPE_STRING,
-      "Printout URI on stdout [-], stderr [+], or a file [anything else]" },
-
-    /* exit status reporting */
-    { "orte_report_child_jobs_separately", '\0', "report-child-jobs-separately", "report-child-jobs-separately", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Return the exit status of the primary job only" },
-
-    /* uri of the dvm, or at least where to get it */
-    { NULL, '\0', "hnp", "hnp", 1,
-      &orte_cmd_line.hnp, OPAL_CMD_LINE_TYPE_STRING,
-      "Specify the URI of the HNP, or the name of the file (specified as file:filename) that contains that info" },
-
-    /* hetero apps */
-    { "orte_hetero_apps", '\0', NULL, "hetero-apps", 0,
-        NULL, OPAL_CMD_LINE_TYPE_BOOL,
-    "Indicates that multiple app_contexts are being provided that are a mix of 32/64 bit binaries" },
-
-    /* select XML output */
-    { "orte_xml_output", '\0', "xml", "xml", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Provide all output in XML format" },
-    { "orte_xml_file", '\0', "xml-file", "xml-file", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Provide all output in XML format to the specified file" },
-
-    /* tag output */
-    { "orte_tag_output", '\0', "tag-output", "tag-output", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Tag all output with [job,rank]" },
-    { "orte_timestamp_output", '\0', "timestamp-output", "timestamp-output", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Timestamp all application process output" },
-    { "orte_output_filename", '\0', "output-filename", "output-filename", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Redirect output from application processes into filename/job/rank/std[out,err,diag]" },
-    { NULL, '\0', "merge-stderr-to-stdout", "merge-stderr-to-stdout", 0,
-      &orte_cmd_line.merge, OPAL_CMD_LINE_TYPE_BOOL,
-      "Merge stderr to stdout for each process"},
-    { "orte_xterm", '\0', "xterm", "xterm", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Create a new xterm window and display output from the specified ranks there" },
-
-    /* select stdin option */
-    { NULL, '\0', "stdin", "stdin", 1,
-      &orte_cmd_line.stdin_target, OPAL_CMD_LINE_TYPE_STRING,
-      "Specify procs to receive stdin [rank, all, none] (default: 0, indicating rank 0)" },
-
-    /* request that argv[0] be indexed */
-    { NULL, '\0', "index-argv-by-rank", "index-argv-by-rank", 0,
-      &orte_cmd_line.index_argv, OPAL_CMD_LINE_TYPE_BOOL,
-      "Uniquely index argv[0] for each process using its rank" },
-
-    /* Specify the launch agent to be used */
-    { "orte_launch_agent", '\0', "launch-agent", "launch-agent", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Command used to start processes on remote nodes (default: orted)" },
-
-    /* Preload the binary on the remote machine */
-    { NULL, 's', NULL, "preload-binary", 0,
-      &orte_cmd_line.preload_binaries, OPAL_CMD_LINE_TYPE_BOOL,
-      "Preload the binary on the remote machine before starting the remote process." },
-
-    /* Preload files on the remote machine */
-    { NULL, '\0', NULL, "preload-files", 1,
-      &orte_cmd_line.preload_files, OPAL_CMD_LINE_TYPE_STRING,
-      "Preload the comma separated list of files to the remote machines current working directory before starting the remote process." },
-
-#if OPAL_ENABLE_FT_CR == 1
-    /* Tell SStore to preload a snapshot before launch */
-    { NULL, '\0', NULL, "sstore-load", 1,
-      &orte_cmd_line.sstore_load, OPAL_CMD_LINE_TYPE_STRING,
-      "Internal Use Only! Tell SStore to preload a snapshot before launch." },
-#endif
-
-    /* Use an appfile */
-    { NULL, '\0', NULL, "app", 1,
-      &orte_cmd_line.appfile, OPAL_CMD_LINE_TYPE_STRING,
-      "Provide an appfile; ignore all other command line options" },
-
-    /* Number of processes; -c, -n, --n, -np, and --np are all
-       synonyms */
-    { NULL, 'c', "np", "np", 1,
-      &orte_cmd_line.num_procs, OPAL_CMD_LINE_TYPE_INT,
-      "Number of processes to run" },
-    { NULL, '\0', "n", "n", 1,
-      &orte_cmd_line.num_procs, OPAL_CMD_LINE_TYPE_INT,
-      "Number of processes to run" },
-
-    /* maximum size of VM - typically used to subdivide an allocation */
-    { "orte_max_vm_size", '\0', "max-vm-size", "max-vm-size", 1,
-      NULL, OPAL_CMD_LINE_TYPE_INT,
-      "Number of processes to run" },
-
-    /* Set a hostfile */
-    { NULL, '\0', "hostfile", "hostfile", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Provide a hostfile" },
-    { NULL, '\0', "machinefile", "machinefile", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Provide a hostfile" },
-    { "orte_default_hostfile", '\0', "default-hostfile", "default-hostfile", 1,
-        NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Provide a default hostfile" },
-    { "opal_if_do_not_resolve", '\0', "do-not-resolve", "do-not-resolve", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Do not attempt to resolve interfaces" },
-
-    /* uri of PMIx publish/lookup server, or at least where to get it */
-    { "pmix_server_uri", '\0', "ompi-server", "ompi-server", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Specify the URI of the publish/lookup server, or the name of the file (specified as file:filename) that contains that info" },
-
-    { "carto_file_path", '\0', "cf", "cartofile", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Provide a cartography file" },
-
-    { "orte_rankfile", '\0', "rf", "rankfile", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Provide a rankfile file" },
-
-    /* Export environment variables; potentially used multiple times,
-       so it does not make sense to set into a variable */
-    { NULL, 'x', NULL, NULL, 1,
-      NULL, OPAL_CMD_LINE_TYPE_NULL,
-      "Export an environment variable, optionally specifying a value (e.g., \"-x foo\" exports the environment variable foo and takes its value from the current environment; \"-x foo=bar\" exports the environment variable name foo and sets its value to \"bar\" in the started processes)" },
-
-      /* Mapping controls */
-    { "rmaps_base_display_map", '\0', "display-map", "display-map", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Display the process map just before launch"},
-    { "rmaps_base_display_devel_map", '\0', "display-devel-map", "display-devel-map", 0,
-       NULL, OPAL_CMD_LINE_TYPE_BOOL,
-       "Display a detailed process map (mostly intended for developers) just before launch"},
-    { "rmaps_base_display_topo_with_map", '\0', "display-topo", "display-topo", 0,
-       NULL, OPAL_CMD_LINE_TYPE_BOOL,
-       "Display the topology as part of the process map (mostly intended for developers) just before launch"},
-    { "rmaps_base_display_diffable_map", '\0', "display-diffable-map", "display-diffable-map", 0,
-       NULL, OPAL_CMD_LINE_TYPE_BOOL,
-       "Display a diffable process map (mostly intended for developers) just before launch"},
-    { NULL, 'H', "host", "host", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "List of hosts to invoke processes on" },
-    { "rmaps_base_no_schedule_local", '\0', "nolocal", "nolocal", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Do not run any MPI applications on the local node" },
-    { "rmaps_base_no_oversubscribe", '\0', "nooversubscribe", "nooversubscribe", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Nodes are not to be oversubscribed, even if the system supports such operation"},
-    { "rmaps_base_oversubscribe", '\0', "oversubscribe", "oversubscribe", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Nodes are allowed to be oversubscribed, even on a managed system, and overloading of processing elements"},
-    { "rmaps_base_cpus_per_rank", '\0', "cpus-per-proc", "cpus-per-proc", 1,
-      NULL, OPAL_CMD_LINE_TYPE_INT,
-      "Number of cpus to use for each process [default=1]" },
-    { "rmaps_base_cpus_per_rank", '\0', "cpus-per-rank", "cpus-per-rank", 1,
-      NULL, OPAL_CMD_LINE_TYPE_INT,
-      "Synonym for cpus-per-proc" },
-
-    /* backward compatiblity */
-    { "rmaps_base_bycore", '\0', "bycore", "bycore", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Whether to map and rank processes round-robin by core" },
-    { "rmaps_base_bynode", '\0', "bynode", "bynode", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Whether to map and rank processes round-robin by node" },
-    { "rmaps_base_byslot", '\0', "byslot", "byslot", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Whether to map and rank processes round-robin by slot" },
-
-    /* Nperxxx options that do not require topology and are always
-     * available - included for backwards compatibility
-     */
-    { "rmaps_ppr_pernode", '\0', "pernode", "pernode", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Launch one process per available node" },
-    { "rmaps_ppr_n_pernode", '\0', "npernode", "npernode", 1,
-        NULL, OPAL_CMD_LINE_TYPE_INT,
-        "Launch n processes per node on all allocated nodes" },
-    { "rmaps_ppr_n_pernode", '\0', "N", NULL, 1,
-        NULL, OPAL_CMD_LINE_TYPE_INT,
-        "Launch n processes per node on all allocated nodes (synonym for npernode)" },
-
-    /* declare hardware threads as independent cpus */
-    { "hwloc_base_use_hwthreads_as_cpus", '\0', "use-hwthread-cpus", "use-hwthread-cpus", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Use hardware threads as independent cpus" },
-
-    /* include npersocket for backwards compatibility */
-    { "rmaps_ppr_n_persocket", '\0', "npersocket", "npersocket", 1,
-      NULL, OPAL_CMD_LINE_TYPE_INT,
-      "Launch n processes per socket on all allocated nodes" },
-
-    /* Mapping options */
-    { "rmaps_base_mapping_policy", '\0', NULL, "map-by", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Mapping Policy [slot | hwthread | core | socket (default) | numa | board | node]" },
-
-      /* Ranking options */
-    { "rmaps_base_ranking_policy", '\0', NULL, "rank-by", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Ranking Policy [slot (default) | hwthread | core | socket | numa | board | node]" },
-
-      /* Binding options */
-    { "hwloc_base_binding_policy", '\0', NULL, "bind-to", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Policy for binding processes. Allowed values: none, hwthread, core, l1cache, l2cache, l3cache, socket, numa, board (\"none\" is the default when oversubscribed, \"core\" is the default when np<=2, and \"socket\" is the default when np>2). Allowed qualifiers: overload-allowed, if-supported" },
-
-    /* backward compatiblity */
-    { "hwloc_base_bind_to_core", '\0', "bind-to-core", "bind-to-core", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Bind processes to cores" },
-    { "hwloc_base_bind_to_socket", '\0', "bind-to-socket", "bind-to-socket", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Bind processes to sockets" },
-
-    { "hwloc_base_report_bindings", '\0', "report-bindings", "report-bindings", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Whether to report process bindings to stderr" },
-
-    /* slot list option */
-    { "hwloc_base_slot_list", '\0', "slot-list", "slot-list", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "List of processor IDs to bind processes to [default=NULL]"},
-
-    /* generalized pattern mapping option */
-    { "rmaps_ppr_pattern", '\0', NULL, "ppr", 1,
-        NULL, OPAL_CMD_LINE_TYPE_STRING,
-        "Comma-separated list of number of processes on a given resource type [default: none]" },
-
-    /* Allocation options */
-    { "orte_display_alloc", '\0', "display-allocation", "display-allocation", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Display the allocation being used by this job"},
-    { "orte_display_devel_alloc", '\0', "display-devel-allocation", "display-devel-allocation", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Display a detailed list (mostly intended for developers) of the allocation being used by this job"},
-    { "hwloc_base_cpu_set", '\0', "cpu-set", "cpu-set", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Comma-separated list of ranges specifying logical cpus allocated to this job [default: none]"},
-
-    /* mpiexec-like arguments */
-    { NULL, '\0', "wdir", "wdir", 1,
-      &orte_cmd_line.wdir, OPAL_CMD_LINE_TYPE_STRING,
-      "Set the working directory of the started processes" },
-    { NULL, '\0', "wd", "wd", 1,
-      &orte_cmd_line.wdir, OPAL_CMD_LINE_TYPE_STRING,
-      "Synonym for --wdir" },
-    { NULL, '\0', "set-cwd-to-session-dir", "set-cwd-to-session-dir", 0,
-      &orte_cmd_line.set_cwd_to_session_dir, OPAL_CMD_LINE_TYPE_BOOL,
-      "Set the working directory of the started processes to their session directory" },
-    { NULL, '\0', "path", "path", 1,
-      &orte_cmd_line.path, OPAL_CMD_LINE_TYPE_STRING,
-      "PATH to be used to look for executables to start processes" },
-
-    /* User-level debugger arguments */
-    { NULL, '\0', "tv", "tv", 0,
-      &orte_cmd_line.debugger, OPAL_CMD_LINE_TYPE_BOOL,
-      "Deprecated backwards compatibility flag; synonym for \"--debug\"" },
-    { NULL, '\0', "debug", "debug", 0,
-      &orte_cmd_line.debugger, OPAL_CMD_LINE_TYPE_BOOL,
-      "Invoke the user-level debugger indicated by the orte_base_user_debugger MCA parameter" },
-    { "orte_base_user_debugger", '\0', "debugger", "debugger", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Sequence of debuggers to search for when \"--debug\" is used" },
-    { "orte_output_debugger_proctable", '\0', "output-proctable", "output-proctable", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Output the debugger proctable after launch" },
-
-    /* OpenRTE arguments */
-    { "orte_debug", 'd', "debug-devel", "debug-devel", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Enable debugging of OpenRTE" },
-
-    { "orte_debug_daemons", '\0', "debug-daemons", "debug-daemons", 0,
-      NULL, OPAL_CMD_LINE_TYPE_INT,
-      "Enable debugging of any OpenRTE daemons used by this application" },
-
-    { "orte_debug_daemons_file", '\0', "debug-daemons-file", "debug-daemons-file", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Enable debugging of any OpenRTE daemons used by this application, storing output in files" },
-
-    { "orte_leave_session_attached", '\0', "leave-session-attached", "leave-session-attached", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Enable debugging of OpenRTE" },
-
-    { "orte_do_not_launch", '\0', "do-not-launch", "do-not-launch", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Perform all necessary operations to prepare to launch the application, but do not actually launch it" },
-
-    { NULL, '\0', NULL, "prefix", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Prefix where Open MPI is installed on remote nodes" },
-    { NULL, '\0', NULL, "noprefix", 0,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Disable automatic --prefix behavior" },
-
-    { "orte_report_launch_progress", '\0', "show-progress", "show-progress", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Output a brief periodic report on launch progress" },
-
-    { "orte_use_regexp", '\0', "use-regexp", "use-regexp", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Use regular expressions for launch" },
-
-    { "orte_report_events", '\0', "report-events", "report-events", 1,
-      NULL, OPAL_CMD_LINE_TYPE_STRING,
-      "Report events to a tool listening at the specified URI" },
-
-    { "orte_enable_recovery", '\0', "enable-recovery", "enable-recovery", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Enable recovery from process failure [Default = disabled]" },
-
-    { "orte_max_restarts", '\0', "max-restarts", "max-restarts", 1,
-      NULL, OPAL_CMD_LINE_TYPE_INT,
-      "Max number of times to restart a failed process" },
-
-    { "orte_hetero_nodes", '\0', NULL, "hetero-nodes", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Nodes in cluster may differ in topology, so send the topology back from each node [Default = false]" },
-
-#if OPAL_ENABLE_CRDEBUG == 1
-    { "opal_cr_enable_crdebug", '\0', "crdebug", "crdebug", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Enable C/R Debugging" },
-#endif
-
-    { NULL, '\0', "disable-recovery", "disable-recovery", 0,
-      &orte_cmd_line.disable_recovery, OPAL_CMD_LINE_TYPE_BOOL,
-      "Disable recovery (resets all recovery options to off)" },
-
-    { "state_novm_select", '\0', "novm", "novm", 0,
-      NULL, OPAL_CMD_LINE_TYPE_BOOL,
-      "Execute without creating an allocation-spanning virtual machine (only start daemons on nodes hosting application procs)" },
-
-    { NULL, '\0', "staged", "staged", 0,
-      &orte_cmd_line.staged_exec, OPAL_CMD_LINE_TYPE_BOOL,
-      "Used staged execution if inadequate resources are present (cannot support MPI jobs)" },
-
-    { NULL, '\0', "allow-run-as-root", "allow-run-as-root", 0,
-      &orte_cmd_line.run_as_root, OPAL_CMD_LINE_TYPE_BOOL,
-      "Allow execution as root (STRONGLY DISCOURAGED)" },
-
-    { NULL, '\0', "personality", "personality", 1,
-      &orte_cmd_line.personality, OPAL_CMD_LINE_TYPE_STRING,
-      "Comma-separated list of programming model, languages, and containers being used (default=\"ompi\")" },
-
-    { NULL, '\0', "dvm", "dvm", 0,
-      &orte_cmd_line.create_dvm, OPAL_CMD_LINE_TYPE_BOOL,
-      "Create a persistent distributed virtual machine (DVM)" },
-
-    /* tell the dvm to terminate */
-    { NULL, '\0', "terminate", "terminate", 0,
-      &orte_cmd_line.terminate_dvm, OPAL_CMD_LINE_TYPE_BOOL,
-      "Terminate the DVM" },
-
-    /* End of list */
-    { NULL, '\0', NULL, NULL, 0,
-      NULL, OPAL_CMD_LINE_TYPE_NULL, NULL }
-};
-
-static int define_cli(opal_cmd_line_t *cli)
-{
-    int i, rc;
-    bool takeus = false;
-
-    opal_output_verbose(1, orte_schizo_base_framework.framework_output,
-                        "%s schizo:ompi: define_cli",
-                        ORTE_NAME_PRINT(ORTE_PROC_MY_NAME));
-
-    /* protect against bozo error */
-    if (NULL == cli) {
-        return ORTE_ERR_BAD_PARAM;
-    }
-
-    if (NULL != orte_schizo_base.personalities) {
-        /* if we aren't included, then ignore us */
-        for (i=0; NULL != orte_schizo_base.personalities[i]; i++) {
-            if (0 == strcmp(orte_schizo_base.personalities[i], "ompi")) {
-                takeus = true;
-                break;
-            }
-        }
-        if (!takeus) {
-            return ORTE_ERR_TAKE_NEXT_OPTION;
-        }
-    }
-
-    /* just add ours to the end */
-    rc = opal_cmd_line_add(cli, cmd_line_init);
-    return rc;
-}
-
-static int parse_cli(int argc, int start, char **argv)
+static int parse_cli(char **personality,
+                     int argc, int start, char **argv)
 {
     int i, j, k;
     bool ignore;
@@ -495,25 +83,15 @@ static int parse_cli(int argc, int start, char **argv)
     };
     bool takeus = false;
 
-    opal_output_verbose(1, orte_schizo_base_framework.framework_output,
-                        "%s schizo:ompi: parse_cli",
-                        ORTE_NAME_PRINT(ORTE_PROC_MY_NAME));
-
-    /* if they gave us a list of personalities,
-     * see if we are included */
-    if (NULL != orte_schizo_base.personalities) {
-        for (i=0; NULL != orte_schizo_base.personalities[i]; i++) {
-            if (0 == strcmp(orte_schizo_base.personalities[i], "ompi")) {
-                takeus = true;
-                break;
-            }
+    /* see if we are included */
+    for (i=0; NULL != personality[i]; i++) {
+        if (0 == strcmp(personality[i], "ompi")) {
+            takeus = true;
+            break;
         }
-        if (!takeus) {
-            return ORTE_ERR_TAKE_NEXT_OPTION;
-        }
-    } else {
-        /* attempt to auto-detect CLI options that
-         * we recognize */
+    }
+    if (!takeus) {
+        return ORTE_ERR_TAKE_NEXT_OPTION;
     }
 
     for (i = 0; i < (argc-start); ++i) {
@@ -584,7 +162,8 @@ static int parse_cli(int argc, int start, char **argv)
     return ORTE_SUCCESS;
 }
 
-static int parse_env(char *path,
+static int parse_env(char **personality,
+                     char *path,
                      opal_cmd_line_t *cmd_line,
                      char **srcenv,
                      char ***dstenv)
@@ -596,21 +175,15 @@ static int parse_env(char *path,
     char **vars;
     bool takeus = false;
 
-    opal_output_verbose(1, orte_schizo_base_framework.framework_output,
-                        "%s schizo:ompi: parse_env",
-                        ORTE_NAME_PRINT(ORTE_PROC_MY_NAME));
-
-    if (NULL != orte_schizo_base.personalities) {
-        /* see if we are included */
-        for (i=0; NULL != orte_schizo_base.personalities[i]; i++) {
-            if (0 == strcmp(orte_schizo_base.personalities[i], "ompi")) {
-                takeus = true;
-                break;
-            }
+    /* see if we are included */
+    for (i=0; NULL != personality[i]; i++) {
+        if (0 == strcmp(personality[i], "ompi")) {
+            takeus = true;
+            break;
         }
-        if (!takeus) {
-            return ORTE_ERR_TAKE_NEXT_OPTION;
-        }
+    }
+    if (!takeus) {
+        return ORTE_ERR_TAKE_NEXT_OPTION;
     }
 
     for (i = 0; NULL != srcenv[i]; ++i) {
@@ -734,21 +307,15 @@ static int setup_fork(orte_job_t *jdata,
     char *num_app_ctx;
     bool takeus = false;
 
-    opal_output_verbose(1, orte_schizo_base_framework.framework_output,
-                        "%s schizo:ompi: setup_fork",
-                        ORTE_NAME_PRINT(ORTE_PROC_MY_NAME));
-
-    if (NULL != orte_schizo_base.personalities) {
     /* see if we are included */
-        for (i=0; NULL != jdata->personality[i]; i++) {
-            if (0 == strcmp(jdata->personality[i], "ompi")) {
-                takeus = true;
-                break;
-            }
+    for (i=0; NULL != jdata->personality[i]; i++) {
+        if (0 == strcmp(jdata->personality[i], "ompi")) {
+            takeus = true;
+            break;
         }
-        if (!takeus) {
-            return ORTE_ERR_TAKE_NEXT_OPTION;
-        }
+    }
+    if (!takeus) {
+        return ORTE_ERR_TAKE_NEXT_OPTION;
     }
 
     /* see if the mapper thinks we are oversubscribed */
@@ -972,21 +539,15 @@ static int setup_child(orte_job_t *jdata,
     int32_t nrestarts=0, *nrptr;
     bool takeus = false;
 
-    opal_output_verbose(1, orte_schizo_base_framework.framework_output,
-                        "%s schizo:ompi: setup_child",
-                        ORTE_NAME_PRINT(ORTE_PROC_MY_NAME));
-
-    if (NULL != orte_schizo_base.personalities) {
-        /* see if we are included */
-        for (i=0; NULL != jdata->personality[i]; i++) {
-            if (0 == strcmp(jdata->personality[i], "ompi")) {
-                takeus = true;
-                break;
-            }
+    /* see if we are included */
+    for (i=0; NULL != jdata->personality[i]; i++) {
+        if (0 == strcmp(jdata->personality[i], "ompi")) {
+            takeus = true;
+            break;
         }
-        if (!takeus) {
-            return ORTE_ERR_TAKE_NEXT_OPTION;
-        }
+    }
+    if (!takeus) {
+        return ORTE_ERR_TAKE_NEXT_OPTION;
     }
 
     /* setup the jobid */
