+++ orte/orted/orted_submit.c
@@ -87,20 +87,16 @@
 #include "orte/mca/rmaps/rmaps_types.h"
 #include "orte/mca/rmaps/base/base.h"
 
+#include "orte/mca/schizo/schizo.h"
 #include "orte/mca/errmgr/errmgr.h"
-#include "orte/mca/plm/base/plm_private.h"
 #include "orte/mca/rml/rml.h"
 #include "orte/mca/rml/base/rml_contact.h"
 #include "orte/mca/routed/routed.h"
-#include "orte/mca/schizo/schizo.h"
-#include "orte/mca/state/state.h"
 
 #include "orte/runtime/runtime.h"
 #include "orte/runtime/orte_globals.h"
 #include "orte/runtime/orte_wait.h"
 #include "orte/runtime/orte_quit.h"
-#include "orte/util/cmd_line.h"
-#include "orte/util/pre_condition_transports.h"
 #include "orte/util/show_help.h"
 
 #include "orted_submit.h"
@@ -116,9 +112,187 @@ static bool want_prefix_by_default = (bool) ORTE_WANT_ORTERUN_PREFIX_BY_DEFAULT;
 static opal_pointer_array_t tool_jobs;
 static opal_cmd_line_t *cmd_line=NULL;
 static bool mycmdline = false;
-int orte_debugger_attach_fd = -1;
-bool orte_debugger_fifo_active=false;
-opal_event_t *orte_debugger_attach=NULL;
+
+static opal_cmd_line_init_t cmd_line_init[] = {
+    /* Various "obvious" options */
+    { NULL, 'h', NULL, "help", 0,
+      &orte_cmd_line.help, OPAL_CMD_LINE_TYPE_BOOL,
+      "This help message" },
+    { NULL, 'V', NULL, "version", 0,
+      &orte_cmd_line.version, OPAL_CMD_LINE_TYPE_BOOL,
+      "Print version and exit" },
+
+    /* tag output */
+    { NULL, '\0', "tag-output", "tag-output", 0,
+      &orte_cmd_line.tag_output, OPAL_CMD_LINE_TYPE_BOOL,
+      "Tag all output with [job,rank]" },
+    { NULL, '\0', "timestamp-output", "timestamp-output", 0,
+      &orte_cmd_line.timestamp_output, OPAL_CMD_LINE_TYPE_BOOL,
+      "Timestamp all application process output" },
+    { NULL, '\0', "output-filename", "output-filename", 1,
+      &orte_cmd_line.output_filename, OPAL_CMD_LINE_TYPE_STRING,
+      "Redirect output from application processes into filename/job/rank/std[out,err,diag]" },
+    { NULL, '\0', "merge-stderr-to-stdout", "merge-stderr-to-stdout", 0,
+      &orte_cmd_line.merge, OPAL_CMD_LINE_TYPE_BOOL,
+      "Merge stderr to stdout for each process"},
+
+    /* select stdin option */
+    { NULL, '\0', "stdin", "stdin", 1,
+      &orte_cmd_line.stdin_target, OPAL_CMD_LINE_TYPE_STRING,
+      "Specify procs to receive stdin [rank, all, none] (default: 0, indicating rank 0)" },
+
+    /* request that argv[0] be indexed */
+    { NULL, '\0', "index-argv-by-rank", "index-argv-by-rank", 0,
+      &orte_cmd_line.index_argv, OPAL_CMD_LINE_TYPE_BOOL,
+      "Uniquely index argv[0] for each process using its rank" },
+
+    /* Preload the binary on the remote machine */
+    { NULL, 's', NULL, "preload-binary", 0,
+      &orte_cmd_line.preload_binaries, OPAL_CMD_LINE_TYPE_BOOL,
+      "Preload the binary on the remote machine before starting the remote process." },
+
+    /* Preload files on the remote machine */
+    { NULL, '\0', NULL, "preload-files", 1,
+      &orte_cmd_line.preload_files, OPAL_CMD_LINE_TYPE_STRING,
+      "Preload the comma separated list of files to the remote machines current working directory before starting the remote process." },
+
+    /* Use an appfile */
+    { NULL, '\0', NULL, "app", 1,
+      &orte_cmd_line.appfile, OPAL_CMD_LINE_TYPE_STRING,
+      "Provide an appfile; ignore all other command line options" },
+
+    /* Number of processes; -c, -n, --n, -np, and --np are all
+       synonyms */
+    { NULL, 'c', "np", "np", 1,
+      &orte_cmd_line.num_procs, OPAL_CMD_LINE_TYPE_INT,
+      "Number of processes to run" },
+    { NULL, '\0', "n", "n", 1,
+      &orte_cmd_line.num_procs, OPAL_CMD_LINE_TYPE_INT,
+      "Number of processes to run" },
+
+    /* uri of the dvm, or at least where to get it */
+    { NULL, '\0', "hnp", "hnp", 1,
+      &orte_cmd_line.hnp, OPAL_CMD_LINE_TYPE_STRING,
+      "Specify the URI of the Open MPI server, or the name of the file (specified as file:filename) that contains that info" },
+
+    /* Set a hostfile */
+    { NULL, '\0', "hostfile", "hostfile", 1,
+      NULL, OPAL_CMD_LINE_TYPE_STRING,
+      "Provide a hostfile" },
+    { NULL, '\0', "machinefile", "machinefile", 1,
+      NULL, OPAL_CMD_LINE_TYPE_STRING,
+      "Provide a hostfile" },
+    { "orte_default_hostfile", '\0', "default-hostfile", "default-hostfile", 1,
+        NULL, OPAL_CMD_LINE_TYPE_STRING,
+    "Provide a default hostfile" },
+    { "opal_if_do_not_resolve", '\0', "do-not-resolve", "do-not-resolve", 0,
+      NULL, OPAL_CMD_LINE_TYPE_BOOL,
+      "Do not attempt to resolve interfaces" },
+
+    /* Export environment variables; potentially used multiple times,
+       so it does not make sense to set into a variable */
+    { NULL, 'x', NULL, NULL, 1,
+      NULL, OPAL_CMD_LINE_TYPE_NULL,
+      "Export an environment variable, optionally specifying a value (e.g., \"-x foo\" exports the environment variable foo and takes its value from the current environment; \"-x foo=bar\" exports the environment variable name foo and sets its value to \"bar\" in the started processes)" },
+
+      /* Mapping controls */
+    { NULL, 'H', "host", "host", 1,
+      NULL, OPAL_CMD_LINE_TYPE_STRING,
+      "List of hosts to invoke processes on" },
+    { NULL, '\0', "nolocal", "nolocal", 0,
+      &orte_cmd_line.nolocal, OPAL_CMD_LINE_TYPE_BOOL,
+      "Do not run any MPI applications on the local node" },
+    { NULL, '\0', "nooversubscribe", "nooversubscribe", 0,
+      &orte_cmd_line.no_oversubscribe, OPAL_CMD_LINE_TYPE_BOOL,
+      "Nodes are not to be oversubscribed, even if the system supports such operation"},
+    { NULL, '\0', "oversubscribe", "oversubscribe", 0,
+      &orte_cmd_line.oversubscribe, OPAL_CMD_LINE_TYPE_BOOL,
+      "Nodes are allowed to be oversubscribed, even on a managed system, and overloading of processing elements"},
+    { NULL, '\0', "cpus-per-proc", "cpus-per-proc", 1,
+      &orte_cmd_line.cpus_per_proc, OPAL_CMD_LINE_TYPE_INT,
+      "Number of cpus to use for each process [default=1]" },
+
+    /* Nperxxx options that do not require topology and are always
+     * available - included for backwards compatibility
+     */
+    { NULL, '\0', "pernode", "pernode", 0,
+      &orte_cmd_line.pernode, OPAL_CMD_LINE_TYPE_BOOL,
+      "Launch one process per available node" },
+    { NULL, '\0', "npernode", "npernode", 1,
+      &orte_cmd_line.npernode, OPAL_CMD_LINE_TYPE_INT,
+      "Launch n processes per node on all allocated nodes" },
+    { NULL, '\0', "N", NULL, 1,
+      &orte_cmd_line.npernode, OPAL_CMD_LINE_TYPE_INT,
+        "Launch n processes per node on all allocated nodes (synonym for npernode)" },
+
+    /* declare hardware threads as independent cpus */
+    { NULL, '\0', "use-hwthread-cpus", "use-hwthread-cpus", 0,
+      &orte_cmd_line.use_hwthreads_as_cpus, OPAL_CMD_LINE_TYPE_BOOL,
+      "Use hardware threads as independent cpus" },
+
+    /* include npersocket for backwards compatibility */
+    { NULL, '\0', "npersocket", "npersocket", 1,
+      &orte_cmd_line.npersocket, OPAL_CMD_LINE_TYPE_INT,
+      "Launch n processes per socket on all allocated nodes" },
+
+    /* Mapping options */
+    { NULL, '\0', NULL, "map-by", 1,
+      &orte_cmd_line.mapping_policy, OPAL_CMD_LINE_TYPE_STRING,
+      "Mapping Policy [slot | hwthread | core | socket (default) | numa | board | node]" },
+
+      /* Ranking options */
+    { NULL, '\0', NULL, "rank-by", 1,
+      &orte_cmd_line.ranking_policy, OPAL_CMD_LINE_TYPE_STRING,
+      "Ranking Policy [slot (default) | hwthread | core | socket | numa | board | node]" },
+
+      /* Binding options */
+    { NULL, '\0', NULL, "bind-to", 1,
+      &orte_cmd_line.binding_policy, OPAL_CMD_LINE_TYPE_STRING,
+      "Policy for binding processes. Allowed values: none, hwthread, core, l1cache, l2cache, l3cache, socket, numa, board (\"none\" is the default when oversubscribed, \"core\" is the default when np<=2, and \"socket\" is the default when np>2). Allowed qualifiers: overload-allowed, if-supported" },
+
+    { NULL, '\0', "report-bindings", "report-bindings", 0,
+      &orte_cmd_line.report_bindings, OPAL_CMD_LINE_TYPE_BOOL,
+      "Whether to report process bindings to stderr" },
+
+    /* slot list option */
+    { NULL, '\0', "slot-list", "slot-list", 1,
+      &orte_cmd_line.slot_list, OPAL_CMD_LINE_TYPE_STRING,
+      "List of processor IDs to bind processes to [default=NULL]"},
+
+    /* mpiexec-like arguments */
+    { NULL, '\0', "wdir", "wdir", 1,
+      &orte_cmd_line.wdir, OPAL_CMD_LINE_TYPE_STRING,
+      "Set the working directory of the started processes" },
+    { NULL, '\0', "wd", "wd", 1,
+      &orte_cmd_line.wdir, OPAL_CMD_LINE_TYPE_STRING,
+      "Synonym for --wdir" },
+    { NULL, '\0', "set-cwd-to-session-dir", "set-cwd-to-session-dir", 0,
+      &orte_cmd_line.set_cwd_to_session_dir, OPAL_CMD_LINE_TYPE_BOOL,
+      "Set the working directory of the started processes to their session directory" },
+    { NULL, '\0', "path", "path", 1,
+      &orte_cmd_line.path, OPAL_CMD_LINE_TYPE_STRING,
+      "PATH to be used to look for executables to start processes" },
+
+    { NULL, '\0', "enable-recovery", "enable-recovery", 0,
+      &orte_cmd_line.enable_recovery, OPAL_CMD_LINE_TYPE_BOOL,
+      "Enable recovery (resets all recovery options to on)" },
+
+    { NULL, '\0', "personality", "personality", 1,
+      &orte_cmd_line.personality, OPAL_CMD_LINE_TYPE_STRING,
+      "Comma-separated list of programming model, languages, and containers being used (default=\"ompi\")" },
+
+    { NULL, 'd', "debug-devel", "debug-devel", 0,
+      &orte_cmd_line.debug, OPAL_CMD_LINE_TYPE_BOOL,
+      "Enable debugging of OpenRTE" },
+
+    { NULL, '\0', "allow-run-as-root", "allow-run-as-root", 0,
+      &orte_cmd_line.run_as_root, OPAL_CMD_LINE_TYPE_BOOL,
+      "Allow execution as root (STRONGLY DISCOURAGED)" },
+
+    /* End of list */
+    { NULL, '\0', NULL, NULL, 0,
+      NULL, OPAL_CMD_LINE_TYPE_NULL, NULL }
+};
 
 /*
  * Local functions
@@ -139,38 +313,6 @@ static void launch_recv(int status, orte_process_name_t* sender,
 static void complete_recv(int status, orte_process_name_t* sender,
                           opal_buffer_t *buffer,
                           orte_rml_tag_t tag, void *cbdata);
-static void attach_debugger(int fd, short event, void *arg);
-static void build_debugger_args(orte_app_context_t *debugger);
-static void open_fifo (void);
-static void run_debugger(char *basename, opal_cmd_line_t *cmd_line,
-                         int argc, char *argv[], int num_procs);
-
-/* instance the standard MPIR interfaces */
-#define MPIR_MAX_PATH_LENGTH 512
-#define MPIR_MAX_ARG_LENGTH 1024
-struct MPIR_PROCDESC *MPIR_proctable = NULL;
-int MPIR_proctable_size = 0;
-volatile int MPIR_being_debugged = 0;
-volatile int MPIR_debug_state = 0;
-int MPIR_i_am_starter = 0;
-int MPIR_partial_attach_ok = 1;
-char MPIR_executable_path[MPIR_MAX_PATH_LENGTH] = {0};
-char MPIR_server_arguments[MPIR_MAX_ARG_LENGTH] = {0};
-volatile int MPIR_forward_output = 0;
-volatile int MPIR_forward_comm = 0;
-char MPIR_attach_fifo[MPIR_MAX_PATH_LENGTH] = {0};
-int MPIR_force_to_main = 0;
-static void orte_debugger_init_before_spawn(orte_job_t *jdata);
-
-ORTE_DECLSPEC void* MPIR_Breakpoint(void);
-
-/*
- * Breakpoint function for parallel debuggers
- */
-void* MPIR_Breakpoint(void)
-{
-    return NULL;
-}
 
 /* local objects */
 typedef struct {
@@ -204,8 +346,6 @@ int orte_submit_init(int argc, char *argv[],
                      opal_cmd_line_t *opts)
 {
     int rc;
-    bool version, help;
-    char *param;
 
     OBJ_CONSTRUCT(&tool_jobs, opal_pointer_array_t);
     opal_pointer_array_init(&tool_jobs, 256, INT_MAX, 128);
@@ -216,8 +356,8 @@ int orte_submit_init(int argc, char *argv[],
 
     /* setup the cmd line only once */
     if (NULL != opts) {
-        /* just add the component-defined ones to the end */
-        if (OPAL_SUCCESS != (rc = orte_schizo.define_cli(opts))) {
+        /* just add ours to the end */
+        if (OPAL_SUCCESS != (rc = opal_cmd_line_add(opts, cmd_line_init))) {
             return rc;
         }
         cmd_line = opts;
@@ -225,46 +365,55 @@ int orte_submit_init(int argc, char *argv[],
     } else {
         /* create our own */
         cmd_line = OBJ_NEW(opal_cmd_line_t);
-        rc = orte_cmd_line_create(cmd_line, argc, argv,
-                                  &environ, &environ,
-                                  &version, &help);
-        if (ORTE_SUCCESS != rc) {
-            OBJ_RELEASE(cmd_line);
-            return rc;
-        }
-        /* print version if requested.  Do this before check for help so
-           that --version --help works as one might expect. */
-        if (version) {
-            char *str, *project_name = NULL;
-            if (0 == strcmp(orte_basename, "mpirun")) {
-                project_name = "Open MPI";
-            } else {
-                project_name = "OpenRTE";
-            }
-            str = opal_info_make_version_str("all",
-                                             OPAL_MAJOR_VERSION, OPAL_MINOR_VERSION,
-                                             OPAL_RELEASE_VERSION,
-                                             OPAL_GREEK_VERSION,
-                                             OPAL_REPO_REV);
-            if (NULL != str) {
-                fprintf(stdout, "%s (%s) %s\n\nReport bugs to %s\n",
-                        orte_basename, project_name, str, PACKAGE_BUGREPORT);
-                free(str);
-            }
-            exit(0);
-        }
+        opal_cmd_line_create(cmd_line, cmd_line_init);
+        mca_base_cmd_line_setup(cmd_line);
         mycmdline = true;
+    }
+
+    /* parse the cmd line - we do this here to get the initial
+     * MCA parameters that might impact our own init */
+    if (OPAL_SUCCESS != (rc = opal_cmd_line_parse(cmd_line, true,
+                                                  argc, argv)) ) {
+        if (OPAL_ERR_SILENT != rc) {
+            fprintf(stderr, "%s: command line error (%s)\n", argv[0],
+                    opal_strerror(rc));
+        }
+        return rc;
+    }
+
+    /* print version if requested.  Do this before check for help so
+       that --version --help works as one might expect. */
+    if (orte_cmd_line.version) {
+        char *str;
+        str = opal_info_make_version_str("all",
+                                         OPAL_MAJOR_VERSION, OPAL_MINOR_VERSION,
+                                         OPAL_RELEASE_VERSION,
+                                         OPAL_GREEK_VERSION,
+                                         OPAL_REPO_REV);
+        if (NULL != str) {
+            fprintf(stdout, "%s %s\n\nReport bugs to %s\n",
+                    orte_basename, str, PACKAGE_BUGREPORT);
+            free(str);
+        }
+        return ORTE_ERR_SILENT;
+    }
+
+   /* process MCA/GMCA parameters */
+    if (OPAL_SUCCESS != (rc = mca_base_cmd_line_process_args(cmd_line, &environ, &environ))) {
+        return rc;
     }
 
     /* Need to initialize OPAL so that install_dirs are filled in */
     if (OPAL_SUCCESS != (rc = opal_init(&argc, &argv))) {
+        OBJ_DESTRUCT(&cmd_line);
         return rc;
     }
 
     /* Check for help request */
-    if (help) {
+    if (orte_cmd_line.help) {
         char *str, *args = NULL;
         char *project_name = NULL;
+
         if (0 == strcmp(orte_basename, "mpirun")) {
             project_name = "Open MPI";
         } else {
@@ -280,157 +429,66 @@ int orte_submit_init(int argc, char *argv[],
             free(str);
         }
         free(args);
+
         /* If someone asks for help, that should be all we do */
         exit(0);
     }
 
-    /* set the flags - if they gave us a -hnp option, then
-     * we are a tool. If not, then we are an HNP */
+    /* if they didn't point us at an HNP, that's an error */
     if (NULL == orte_cmd_line.hnp) {
-        orte_process_info.proc_type = ORTE_PROC_HNP;
-    } else {
-        orte_process_info.proc_type = ORTE_PROC_TOOL;
+        fprintf(stderr, "%s submit: required option --hnp not provided\n", orte_basename);
+        return ORTE_ERROR;
     }
 
-    if (ORTE_PROC_IS_TOOL) {
-        if (0 == strncasecmp(orte_cmd_line.hnp, "file", strlen("file"))) {
-            char input[1024], *filename;
-            FILE *fp;
+    if (0 == strncasecmp(orte_cmd_line.hnp, "file", strlen("file"))) {
+        char input[1024], *filename;
+        FILE *fp;
 
-            /* it is a file - get the filename */
-            filename = strchr(orte_cmd_line.hnp, ':');
-            if (NULL == filename) {
-                /* filename is not correctly formatted */
-                orte_show_help("help-orte-top.txt", "orte-top:hnp-filename-bad", true, "uri", orte_cmd_line.hnp);
-                exit(1);
-            }
-            ++filename; /* space past the : */
+        /* it is a file - get the filename */
+        filename = strchr(orte_cmd_line.hnp, ':');
+        if (NULL == filename) {
+            /* filename is not correctly formatted */
+            orte_show_help("help-orte-top.txt", "orte-top:hnp-filename-bad", true, "uri", orte_cmd_line.hnp);
+            exit(1);
+        }
+        ++filename; /* space past the : */
 
-            if (0 >= strlen(filename)) {
-                /* they forgot to give us the name! */
-                orte_show_help("help-orte-top.txt", "orte-top:hnp-filename-bad", true, "uri", orte_cmd_line.hnp);
-                exit(1);
-            }
+        if (0 >= strlen(filename)) {
+            /* they forgot to give us the name! */
+            orte_show_help("help-orte-top.txt", "orte-top:hnp-filename-bad", true, "uri", orte_cmd_line.hnp);
+            exit(1);
+        }
 
-            /* open the file and extract the uri */
-            fp = fopen(filename, "r");
-            if (NULL == fp) { /* can't find or read file! */
-                orte_show_help("help-orte-top.txt", "orte-top:hnp-filename-access", true, orte_cmd_line.hnp);
-                exit(1);
-            }
-            /* initialize the input to NULLs to ensure any input
-             * string is NULL-terminated */
-            memset(input, 0, 1024);
-            if (NULL == fgets(input, 1024, fp)) {
-                /* something malformed about file */
-                fclose(fp);
-                orte_show_help("help-orte-top.txt", "orte-top:hnp-file-bad", true, orte_cmd_line.hnp);
-                exit(1);
-            }
+        /* open the file and extract the uri */
+        fp = fopen(filename, "r");
+        if (NULL == fp) { /* can't find or read file! */
+            orte_show_help("help-orte-top.txt", "orte-top:hnp-filename-access", true, orte_cmd_line.hnp);
+            exit(1);
+        }
+        /* initialize the input to NULLs to ensure any input
+         * string is NULL-terminated */
+        memset(input, 0, 1024);
+        if (NULL == fgets(input, 1024, fp)) {
+            /* something malformed about file */
             fclose(fp);
-            input[strlen(input)-1] = '\0';  /* remove newline */
-            /* construct the target hnp info */
-            opal_setenv(OPAL_MCA_PREFIX"orte_hnp_uri", input, true, &environ);
-        } else {
-            /* should just be the uri itself - construct the target hnp info */
-            opal_setenv(OPAL_MCA_PREFIX"orte_hnp_uri", orte_cmd_line.hnp, true, &environ);
+            orte_show_help("help-orte-top.txt", "orte-top:hnp-file-bad", true, orte_cmd_line.hnp);
+            exit(1);
         }
-        /* we are never allowed to operate as a distributed tool,
-         * so insist on the ess/tool component */
-        opal_setenv(OPAL_MCA_PREFIX"ess", "tool", true, &environ);
+        fclose(fp);
+        input[strlen(input)-1] = '\0';  /* remove newline */
+        /* construct the target hnp info */
+        opal_setenv(OPAL_MCA_PREFIX"orte_hnp_uri", input, true, &environ);
     } else {
-        /* may look strange, but the way we handle prefix is a little weird
-         * and probably needs to be addressed more fully at some future point.
-         * For now, we have a conflict between app_files and cmd line usage.
-         * Since app_files are used by the C/R system, we will make an
-         * adjustment here to avoid perturbing that system.
-         *
-         * We cannot just have the cmd line parser place any found value
-         * in the global struct as the app_file parser would replace it.
-         * So handle this specific cmd line option manually.
-         */
-        orte_cmd_line.prefix = NULL;
-        orte_cmd_line.path_to_mpirun = NULL;
-        if (opal_cmd_line_is_taken(cmd_line, "prefix") ||
-            '/' == argv[0][0] || want_prefix_by_default) {
-            size_t param_len;
-            if ('/' == argv[0][0]) {
-                char* tmp_basename = NULL;
-                /* If they specified an absolute path, strip off the
-                   /bin/<exec_name>" and leave just the prefix */
-                orte_cmd_line.path_to_mpirun = opal_dirname(argv[0]);
-                /* Quick sanity check to ensure we got
-                   something/bin/<exec_name> and that the installation
-                   tree is at least more or less what we expect it to
-                   be */
-                tmp_basename = opal_basename(orte_cmd_line.path_to_mpirun);
-                if (0 == strcmp("bin", tmp_basename)) {
-                    char* tmp = orte_cmd_line.path_to_mpirun;
-                    orte_cmd_line.path_to_mpirun = opal_dirname(tmp);
-                    free(tmp);
-                } else {
-                    free(orte_cmd_line.path_to_mpirun);
-                    orte_cmd_line.path_to_mpirun = NULL;
-                }
-                free(tmp_basename);
-            }
-            /* if both are given, check to see if they match */
-            if (opal_cmd_line_is_taken(cmd_line, "prefix") && NULL != orte_cmd_line.path_to_mpirun) {
-                char *tmp_basename;
-                /* if they don't match, then that merits a warning */
-                param = strdup(opal_cmd_line_get_param(cmd_line, "prefix", 0, 0));
-                /* ensure we strip any trailing '/' */
-                if (0 == strcmp(OPAL_PATH_SEP, &(param[strlen(param)-1]))) {
-                    param[strlen(param)-1] = '\0';
-                }
-                tmp_basename = strdup(orte_cmd_line.path_to_mpirun);
-                if (0 == strcmp(OPAL_PATH_SEP, &(tmp_basename[strlen(tmp_basename)-1]))) {
-                    tmp_basename[strlen(tmp_basename)-1] = '\0';
-                }
-                if (0 != strcmp(param, tmp_basename)) {
-                    orte_show_help("help-orterun.txt", "orterun:double-prefix",
-                                   true, orte_basename, orte_basename,
-                                   param, tmp_basename, orte_basename);
-                    /* use the prefix over the path-to-mpirun so that
-                     * people can specify the backend prefix as different
-                     * from the local one
-                     */
-                    free(orte_cmd_line.path_to_mpirun);
-                    orte_cmd_line.path_to_mpirun = NULL;
-                }
-                free(tmp_basename);
-            } else if (NULL != orte_cmd_line.path_to_mpirun) {
-                param = strdup(orte_cmd_line.path_to_mpirun);
-            } else if (opal_cmd_line_is_taken(cmd_line, "prefix")){
-                /* must be --prefix alone */
-                param = strdup(opal_cmd_line_get_param(cmd_line, "prefix", 0, 0));
-            } else {
-                /* --enable-orterun-prefix-default was given to orterun */
-                param = strdup(opal_install_dirs.prefix);
-            }
-
-            if (NULL != param) {
-                /* "Parse" the param, aka remove superfluous path_sep. */
-                param_len = strlen(param);
-                while (0 == strcmp (OPAL_PATH_SEP, &(param[param_len-1]))) {
-                    param[param_len-1] = '\0';
-                    param_len--;
-                    if (0 == param_len) {
-                        orte_show_help("help-orterun.txt", "orterun:empty-prefix",
-                                       true, orte_basename, orte_basename);
-                        free(param);
-                        return ORTE_ERR_FATAL;
-                    }
-                }
-
-                orte_cmd_line.prefix = param;
-            }
-            want_prefix_by_default = true;
-        }
+        /* should just be the uri itself - construct the target hnp info */
+        opal_setenv(OPAL_MCA_PREFIX"orte_hnp_uri", orte_cmd_line.hnp, true, &environ);
     }
 
     /* Setup MCA params */
     orte_register_params();
+
+    /* we are never allowed to operate as a distributed tool,
+     * so insist on the ess/tool component */
+    opal_setenv(OPAL_MCA_PREFIX"ess", "tool", true, &environ);
 
     if (orte_cmd_line.debug) {
         orte_devel_level_output = true;
@@ -442,8 +500,7 @@ int orte_submit_init(int argc, char *argv[],
       * up incorrect infrastructure that only a singleton would
       * require
       */
-    if (ORTE_SUCCESS != (rc = orte_init(&argc, &argv,
-                                        orte_process_info.proc_type))) {
+    if (ORTE_SUCCESS != (rc = orte_init(&argc, &argv, ORTE_PROC_TOOL))) {
         /* cannot call ORTE_ERROR_LOG as it could be the errmgr
          * never got loaded!
          */
@@ -454,43 +511,32 @@ int orte_submit_init(int argc, char *argv[],
      */
     opal_finalize();
 
-    /* clear params from the environment so our children
-     * don't pick them up */
+    /* clear the ess param from the environment so our children
+     * don't pick it up */
     opal_unsetenv(OPAL_MCA_PREFIX"ess", &environ);
-    opal_unsetenv(OPAL_MCA_PREFIX"pmix", &environ);
 
-    if (ORTE_PROC_IS_TOOL) {
-        /* set the info in our contact table */
-        orte_rml.set_contact_info(orte_process_info.my_hnp_uri);
-        /* extract the name */
-        if (ORTE_SUCCESS != orte_rml_base_parse_uris(orte_process_info.my_hnp_uri, ORTE_PROC_MY_HNP, NULL)) {
-            orte_show_help("help-orte-top.txt", "orte-top:hnp-uri-bad", true, orte_process_info.my_hnp_uri);
-            exit(1);
-        }
-        /* set the route to be direct */
-        if (ORTE_SUCCESS != orte_routed.update_route(ORTE_PROC_MY_HNP, ORTE_PROC_MY_HNP)) {
-            orte_show_help("help-orte-top.txt", "orte-top:hnp-uri-bad", true, orte_process_info.my_hnp_uri);
-            orte_finalize();
-            exit(1);
-        }
-
-        /* set the target hnp as our lifeline so we will terminate if it exits */
-        orte_routed.set_lifeline(ORTE_PROC_MY_HNP);
-
-        /* setup to listen for HNP response to my commands */
-        orte_rml.recv_buffer_nb(ORTE_NAME_WILDCARD, ORTE_RML_TAG_NOTIFY_COMPLETE,
-                                ORTE_RML_PERSISTENT, complete_recv, NULL);
-        orte_rml.recv_buffer_nb(ORTE_NAME_WILDCARD, ORTE_RML_TAG_LAUNCH_RESP,
-                                ORTE_RML_PERSISTENT, launch_recv, NULL);
-    } else {
-        /* save the environment for launch purposes. This MUST be
-         * done so that we can pass it to any local procs we
-         * spawn - otherwise, those local procs won't see any
-         * non-MCA envars were set in the enviro prior to calling
-         * orterun
-         */
-        orte_launch_environ = opal_argv_copy(environ);
+    /* set the info in our contact table */
+    orte_rml.set_contact_info(orte_process_info.my_hnp_uri);
+    /* extract the name */
+    if (ORTE_SUCCESS != orte_rml_base_parse_uris(orte_process_info.my_hnp_uri, ORTE_PROC_MY_HNP, NULL)) {
+        orte_show_help("help-orte-top.txt", "orte-top:hnp-uri-bad", true, orte_process_info.my_hnp_uri);
+        exit(1);
     }
+    /* set the route to be direct */
+    if (ORTE_SUCCESS != orte_routed.update_route(ORTE_PROC_MY_HNP, ORTE_PROC_MY_HNP)) {
+        orte_show_help("help-orte-top.txt", "orte-top:hnp-uri-bad", true, orte_process_info.my_hnp_uri);
+        orte_finalize();
+        exit(1);
+    }
+
+    /* set the target hnp as our lifeline so we will terminate if it exits */
+    orte_routed.set_lifeline(ORTE_PROC_MY_HNP);
+
+    /* setup to listen for HNP response to my commands */
+    orte_rml.recv_buffer_nb(ORTE_NAME_WILDCARD, ORTE_RML_TAG_NOTIFY_COMPLETE,
+                            ORTE_RML_PERSISTENT, complete_recv, NULL);
+    orte_rml.recv_buffer_nb(ORTE_NAME_WILDCARD, ORTE_RML_TAG_LAUNCH_RESP,
+                            ORTE_RML_PERSISTENT, launch_recv, NULL);
 
     return ORTE_SUCCESS;
 }
@@ -587,18 +633,9 @@ int orte_submit_job(char *argv[], int *index,
     int rc;
     orte_daemon_cmd_flag_t cmd = ORTE_DAEMON_SPAWN_JOB_CMD;
     char *param;
-    orte_job_t *jdata = NULL, *daemons;
-    orte_app_context_t *app, *dapp;
+    orte_job_t *jdata = NULL;
     trackr_t *trk;
     int argc;
-
-    /* bozo check - we don't allow recursive calls of submit */
-    if (NULL != getenv("OMPI_UNIVERSE_SIZE")) {
-        fprintf(stderr, "\n\n**********************************************************\n\n");
-        fprintf(stderr, "%s does not support recursive calls\n", orte_basename);
-        fprintf(stderr, "\n**********************************************************\n");
-        return ORTE_ERR_FATAL;
-    }
 
     /* reset the globals every time thru as the argv
      * will modify them */
@@ -619,6 +656,13 @@ int orte_submit_job(char *argv[], int *index,
     /* Check for some "global" command line params */
     parse_globals(argc, argv, cmd_line);
 
+    /* default our personality to OMPI */
+    if (NULL == orte_cmd_line.personality) {
+        opal_argv_append_nosize(&orte_cmd_line.personalities, "ompi");
+    } else {
+        orte_cmd_line.personalities = opal_argv_split(orte_cmd_line.personality, ',');
+    }
+
     /* create a new job object to hold the info for this one - the
      * jobid field will be filled in by the PLM when the job is
      * launched
@@ -630,6 +674,7 @@ int orte_submit_job(char *argv[], int *index,
          */
         return ORTE_ERR_OUT_OF_RESOURCE;
     }
+    jdata->personality = opal_argv_copy(orte_cmd_line.personalities);
     trk = OBJ_NEW(trackr_t);
     trk->jdata = jdata;
     trk->launch_cb = launch_cb;
@@ -641,6 +686,12 @@ int orte_submit_job(char *argv[], int *index,
 
     /* pass our tracker ID */
     orte_set_attribute(&jdata->attributes, ORTE_JOB_ROOM_NUM, ORTE_ATTR_GLOBAL, &trk->index, OPAL_INT);
+    /* flag that we are using the DVM */
+    orte_set_attribute(&jdata->attributes, ORTE_JOB_DVM_JOB, ORTE_ATTR_GLOBAL, NULL, OPAL_BOOL);
+    /* flag that the allocation is static - i.e., the DVM is not allowed
+     * to be adjusted once started, and all unused nodes are to be
+     * removed from the node pool */
+    orte_set_attribute(&jdata->attributes, ORTE_JOB_FIXED_DVM, ORTE_ATTR_GLOBAL, NULL, OPAL_BOOL);
 
     /* check for stdout/err directives */
     /* if we were asked to tag output, mark it so */
@@ -679,15 +730,6 @@ int orte_submit_job(char *argv[], int *index,
 
     /* Parse each app, adding it to the job object */
     parse_locals(jdata, argc, argv);
-
-    if (0 == jdata->num_apps) {
-        /* This should never happen -- this case should be caught in
-           create_app(), but let's just double check... */
-        orte_show_help("help-orterun.txt", "orterun:nothing-to-do",
-                       true, orte_basename);
-        ORTE_UPDATE_EXIT_STATUS(ORTE_ERROR_DEFAULT_EXIT_CODE);
-        return ORTE_ERR_FATAL;
-    }
 
     /* create the map object to communicate policies */
     jdata->map = OBJ_NEW(orte_job_map_t);
@@ -741,17 +783,12 @@ int orte_submit_job(char *argv[], int *index,
         orte_set_attribute(&jdata->attributes, ORTE_JOB_SLOT_LIST, ORTE_ATTR_GLOBAL, orte_cmd_line.slot_list, OPAL_STRING);
     }
 
-    /* if recovery was disabled on the cmd line, do so */
-    if (orte_cmd_line.enable_recovery) {
-        ORTE_FLAG_SET(jdata, ORTE_JOB_FLAG_RECOVERABLE);
-    }
-
-    /* check for suicide test directives */
-    if (NULL != getenv("ORTE_TEST_HNP_SUICIDE") ||
-        NULL != getenv("ORTE_TEST_ORTED_SUICIDE")) {
-        /* don't forward IO from this process so we can
-         * see any debug after daemon termination */
-        ORTE_FLAG_UNSET(jdata, ORTE_JOB_FLAG_FORWARD_OUTPUT);
+    if (0 == jdata->num_apps) {
+        /* This should never happen -- this case should be caught in
+           create_app(), but let's just double check... */
+        orte_show_help("help-orterun.txt", "orterun:nothing-to-do",
+                       true, orte_basename);
+        return ORTE_ERROR_DEFAULT_EXIT_CODE;
     }
 
     /* check for a job timeout specification, to be provided in seconds
@@ -771,137 +808,31 @@ int orte_submit_job(char *argv[], int *index,
         opal_event_evtimer_add(orte_mpiexec_timeout->ev, &orte_mpiexec_timeout->tv);
     }
 
-    if (ORTE_PROC_IS_HNP) {
-        /* get the daemon job object */
-        daemons = orte_get_job_data_object(ORTE_PROC_MY_NAME->jobid);
-
-        /* check for request to report uri */
-        if (NULL != orte_cmd_line.report_uri) {
-            FILE *fp;
-            char *rml_uri;
-            rml_uri = orte_rml.get_contact_info();
-            if (0 == strcmp(orte_cmd_line.report_uri, "-")) {
-                /* if '-', then output to stdout */
-                printf("%s\n",  (NULL == rml_uri) ? "NULL" : rml_uri);
-            } else if (0 == strcmp(orte_cmd_line.report_uri, "+")) {
-                /* if '+', output to stderr */
-                fprintf(stderr, "%s\n",  (NULL == rml_uri) ? "NULL" : rml_uri);
-            } else {
-                fp = fopen(orte_cmd_line.report_uri, "w");
-                if (NULL == fp) {
-                    orte_show_help("help-orterun.txt", "orterun:write_file", false,
-                                   orte_basename, "uri", orte_cmd_line.report_uri);
-                    exit(0);
-                }
-                fprintf(fp, "%s\n", (NULL == rml_uri) ? "NULL" : rml_uri);
-                fclose(fp);
-            }
-            if (NULL != rml_uri) {
-                free(rml_uri);
-            }
-        }
-        /* If we have a prefix, then modify the PATH and
-           LD_LIBRARY_PATH environment variables in our copy. This
-           will ensure that any locally-spawned children will
-           have our executables and libraries in their path
-
-           For now, default to the prefix_dir provided in the first app_context.
-           Since there always MUST be at least one app_context, we are safe in
-           doing this.
-        */
-        param = NULL;
-        if (NULL != (app = (orte_app_context_t*)opal_pointer_array_get_item(jdata->apps, 0)) &&
-            orte_get_attribute(&app->attributes, ORTE_APP_PREFIX_DIR, (void**)&param, OPAL_STRING)) {
-            char *oldenv, *newenv, *lib_base, *bin_base;
-
-            /* copy the prefix into the daemon job so that any launcher
-             * can find the orteds when we launch the virtual machine
-             */
-            if (NULL == (dapp = (orte_app_context_t*)opal_pointer_array_get_item(daemons->apps, 0))) {
-                /* that's an error in the ess */
-                ORTE_ERROR_LOG(ORTE_ERR_NOT_FOUND);
-                return ORTE_ERR_NOT_FOUND;
-            }
-            orte_set_attribute(&dapp->attributes, ORTE_APP_PREFIX_DIR, ORTE_ATTR_LOCAL, param, OPAL_STRING);
-
-            lib_base = opal_basename(opal_install_dirs.libdir);
-            bin_base = opal_basename(opal_install_dirs.bindir);
-
-            /* Reset PATH */
-            newenv = opal_os_path( false, param, bin_base, NULL );
-            oldenv = getenv("PATH");
-            if (NULL != oldenv) {
-                char *temp;
-                asprintf(&temp, "%s:%s", newenv, oldenv );
-                free( newenv );
-                newenv = temp;
-            }
-            opal_setenv("PATH", newenv, true, &orte_launch_environ);
-            if (orte_debug_flag) {
-                opal_output(0, "%s: reset PATH: %s", orte_basename, newenv);
-            }
-            free(newenv);
-            free(bin_base);
-
-            /* Reset LD_LIBRARY_PATH */
-            newenv = opal_os_path( false, param, lib_base, NULL );
-            oldenv = getenv("LD_LIBRARY_PATH");
-            if (NULL != oldenv) {
-                char* temp;
-                asprintf(&temp, "%s:%s", newenv, oldenv);
-                free(newenv);
-                newenv = temp;
-            }
-            opal_setenv("LD_LIBRARY_PATH", newenv, true, &orte_launch_environ);
-            if (orte_debug_flag) {
-                opal_output(0, "%s: reset LD_LIBRARY_PATH: %s",
-                            orte_basename, newenv);
-            }
-            free(newenv);
-            free(lib_base);
-            free(param);
-        }
-
-        /* pre-condition any network transports that require it */
-        if (ORTE_SUCCESS != (rc = orte_pre_condition_transports(jdata))) {
-            ORTE_ERROR_LOG(rc);
-            orte_show_help("help-orterun.txt", "orterun:precondition", false,
-                           orte_basename, NULL, NULL, rc);
-            ORTE_UPDATE_EXIT_STATUS(ORTE_ERROR_DEFAULT_EXIT_CODE);
-            return rc;
-        }
-        /* setup for debugging */
-        orte_debugger_init_before_spawn(jdata);
-
-        rc = orte_plm.spawn(jdata);
-    } else {
-        /* flag that we are using the DVM */
-        orte_set_attribute(&jdata->attributes, ORTE_JOB_DVM_JOB, ORTE_ATTR_GLOBAL, NULL, OPAL_BOOL);
-        /* flag that the allocation is static - i.e., the DVM is not allowed
-         * to be adjusted once started, and all unused nodes are to be
-         * removed from the node pool */
-        orte_set_attribute(&jdata->attributes, ORTE_JOB_FIXED_DVM, ORTE_ATTR_GLOBAL, NULL, OPAL_BOOL);
-        // pack the ORTE_DAEMON_SPAWN_JOB_CMD command and job object and send to HNP at tag ORTE_RML_TAG_DAEMON
-        req = OBJ_NEW(opal_buffer_t);
-        if (OPAL_SUCCESS != (rc = opal_dss.pack(req, &cmd, 1, ORTE_DAEMON_CMD))) {
-            ORTE_ERROR_LOG(rc);
-            return rc;
-        }
-        if (OPAL_SUCCESS != (rc = opal_dss.pack(req, &jdata, 1, ORTE_JOB))) {
-            ORTE_ERROR_LOG(rc);
-            return rc;
-        }
-        if (OPAL_SUCCESS != (rc = opal_dss.pack(req, &trk->index, 1, OPAL_INT))) {
-            ORTE_ERROR_LOG(rc);
-            return rc;
-        }
-        orte_rml.send_buffer_nb(ORTE_PROC_MY_HNP, req, ORTE_RML_TAG_DAEMON, orte_rml_send_callback, NULL);
-
-        /* Inform the caller of the tracker index if they passed a index pointer */
-        if (NULL != index) {
-            *index = trk->index;
-        }
+    /* if recovery was disabled on the cmd line, do so */
+    if (orte_cmd_line.enable_recovery) {
+        ORTE_FLAG_SET(jdata, ORTE_JOB_FLAG_RECOVERABLE);
     }
+
+    // pack the ORTE_DAEMON_SPAWN_JOB_CMD command and job object and send to HNP at tag ORTE_RML_TAG_DAEMON
+    req = OBJ_NEW(opal_buffer_t);
+    if (OPAL_SUCCESS != (rc = opal_dss.pack(req, &cmd, 1, ORTE_DAEMON_CMD))) {
+        ORTE_ERROR_LOG(rc);
+        return rc;
+    }
+    if (OPAL_SUCCESS != (rc = opal_dss.pack(req, &jdata, 1, ORTE_JOB))) {
+        ORTE_ERROR_LOG(rc);
+        return rc;
+    }
+    if (OPAL_SUCCESS != (rc = opal_dss.pack(req, &trk->index, 1, OPAL_INT))) {
+        ORTE_ERROR_LOG(rc);
+        return rc;
+    }
+
+    orte_rml.send_buffer_nb(ORTE_PROC_MY_HNP, req, ORTE_RML_TAG_DAEMON, orte_rml_send_callback, NULL);
+
+    /* Inform the caller of the tracker index if they passed a index pointer */
+    if (NULL != index)
+        *index = trk->index;
 
     return ORTE_SUCCESS;
 
@@ -957,12 +888,6 @@ static int parse_globals(int argc, char* argv[], opal_cmd_line_t *cmd_line)
             fprintf(fp, "%d\n", (int)getpid());
             fclose(fp);
         }
-    }
-
-    /* Do we want a user-level debugger? */
-
-    if (orte_cmd_line.debugger) {
-        run_debugger(orte_basename, cmd_line, argc, argv, orte_cmd_line.num_procs);
     }
 
     return ORTE_SUCCESS;
@@ -1146,7 +1071,7 @@ static int create_app(int argc, char* argv[],
                       orte_app_context_t **app_ptr,
                       bool *made_app, char ***app_env)
 {
-    opal_cmd_line_t app_cmd_line;
+    opal_cmd_line_t cmd_line;
     char cwd[OPAL_PATH_MAX];
     int i, j, count, rc;
     char *param, *value;
@@ -1168,28 +1093,31 @@ static int create_app(int argc, char* argv[],
      * Only pick up '-mca foo bar' on this pass.
      */
     if (NULL != orte_cmd_line.appfile) {
-        if (ORTE_SUCCESS != (rc = orte_schizo.parse_cli(argc, 0, argv))) {
+        if (ORTE_SUCCESS != (rc = orte_schizo.parse_cli(orte_cmd_line.personalities, argc, 0, argv))) {
             goto cleanup;
         }
     }
 
     /* Parse application command line options. */
     init_globals();
-    OBJ_CONSTRUCT(&app_cmd_line, opal_cmd_line_t);
-    rc = orte_cmd_line_create(&app_cmd_line, argc, argv,
-                              app_env, &global_mca_env,
-                              NULL, NULL);
+    opal_cmd_line_create(&cmd_line, cmd_line_init);
+    mca_base_cmd_line_setup(&cmd_line);
     cmd_line_made = true;
+    rc = opal_cmd_line_parse(&cmd_line, true, argc, argv);
+    if (ORTE_SUCCESS != rc) {
+        goto cleanup;
+    }
+    mca_base_cmd_line_process_args(&cmd_line, app_env, &global_mca_env);
 
     /* Is there an appfile in here? */
     if (NULL != orte_cmd_line.appfile) {
-        OBJ_DESTRUCT(&app_cmd_line);
+        OBJ_DESTRUCT(&cmd_line);
         return parse_appfile(jdata, strdup(orte_cmd_line.appfile), app_env);
     }
 
     /* Setup application context */
     app = OBJ_NEW(orte_app_context_t);
-    opal_cmd_line_get_tail(&app_cmd_line, &count, &app->argv);
+    opal_cmd_line_get_tail(&cmd_line, &count, &app->argv);
 
     /* See if we have anything left */
     if (0 == count) {
@@ -1206,15 +1134,16 @@ static int create_app(int argc, char* argv[],
      *   mpirun -np 2 -mca foo bar ./my-app -mca bip bop
      * We want to pick up '-mca foo bar' but not '-mca bip bop'
      */
-    if (ORTE_SUCCESS != (rc = orte_schizo.parse_cli(argc, count, argv))) {
+    if (ORTE_SUCCESS != (rc = orte_schizo.parse_cli(orte_cmd_line.personalities, argc, count, argv))) {
         goto cleanup;
     }
 
     /* Grab all MCA environment variables */
 
     app->env = opal_argv_copy(*app_env);
-    if (ORTE_SUCCESS != (rc = orte_schizo.parse_env(orte_cmd_line.path,
-                                                    &app_cmd_line,
+    if (ORTE_SUCCESS != (rc = orte_schizo.parse_env(orte_cmd_line.personalities,
+                                                    orte_cmd_line.path,
+                                                    &cmd_line,
                                                     environ, &app->env))) {
         goto cleanup;
     }
@@ -1258,20 +1187,20 @@ static int create_app(int argc, char* argv[],
         /* Check to see if the user explicitly wanted to disable automatic
            --prefix behavior */
 
-        if (opal_cmd_line_is_taken(&app_cmd_line, "noprefix")) {
+        if (opal_cmd_line_is_taken(&cmd_line, "noprefix")) {
             want_prefix_by_default = false;
         }
 
         /* Did the user specify a prefix, or want prefix by default? */
-        if (opal_cmd_line_is_taken(&app_cmd_line, "prefix") || want_prefix_by_default) {
+        if (opal_cmd_line_is_taken(&cmd_line, "prefix") || want_prefix_by_default) {
             size_t param_len;
             /* if both the prefix was given and we have a prefix
              * given above, check to see if they match
              */
-            if (opal_cmd_line_is_taken(&app_cmd_line, "prefix") &&
+            if (opal_cmd_line_is_taken(&cmd_line, "prefix") &&
                 NULL != orte_cmd_line.prefix) {
                 /* if they don't match, then that merits a warning */
-                param = strdup(opal_cmd_line_get_param(&app_cmd_line, "prefix", 0, 0));
+                param = strdup(opal_cmd_line_get_param(&cmd_line, "prefix", 0, 0));
                 /* ensure we strip any trailing '/' */
                 if (0 == strcmp(OPAL_PATH_SEP, &(param[strlen(param)-1]))) {
                     param[strlen(param)-1] = '\0';
@@ -1292,9 +1221,9 @@ static int create_app(int argc, char* argv[],
                 free(value);
             } else if (NULL != orte_cmd_line.prefix) {
                 param = strdup(orte_cmd_line.prefix);
-            } else if (opal_cmd_line_is_taken(&app_cmd_line, "prefix")){
+            } else if (opal_cmd_line_is_taken(&cmd_line, "prefix")){
                 /* must be --prefix alone */
-                param = strdup(opal_cmd_line_get_param(&app_cmd_line, "prefix", 0, 0));
+                param = strdup(opal_cmd_line_get_param(&cmd_line, "prefix", 0, 0));
             } else {
                 /* --enable-orterun-prefix-default was given to orterun */
                 param = strdup(opal_install_dirs.prefix);
@@ -1323,32 +1252,32 @@ static int create_app(int argc, char* argv[],
      * hostfile and machine file.
      * We can only deal with one hostfile per app context, otherwise give an error.
      */
-    if (0 < (j = opal_cmd_line_get_ninsts(&app_cmd_line, "hostfile"))) {
+    if (0 < (j = opal_cmd_line_get_ninsts(&cmd_line, "hostfile"))) {
         if(1 < j) {
             orte_show_help("help-orterun.txt", "orterun:multiple-hostfiles",
                            true, orte_basename, NULL);
             return ORTE_ERR_FATAL;
         } else {
-            value = opal_cmd_line_get_param(&app_cmd_line, "hostfile", 0, 0);
+            value = opal_cmd_line_get_param(&cmd_line, "hostfile", 0, 0);
             orte_set_attribute(&app->attributes, ORTE_APP_HOSTFILE, ORTE_ATTR_GLOBAL, value, OPAL_STRING);
         }
     }
-    if (0 < (j = opal_cmd_line_get_ninsts(&app_cmd_line, "machinefile"))) {
+    if (0 < (j = opal_cmd_line_get_ninsts(&cmd_line, "machinefile"))) {
         if(1 < j || orte_get_attribute(&app->attributes, ORTE_APP_HOSTFILE, NULL, OPAL_STRING)) {
             orte_show_help("help-orterun.txt", "orterun:multiple-hostfiles",
                            true, orte_basename, NULL);
             return ORTE_ERR_FATAL;
         } else {
-            value = opal_cmd_line_get_param(&app_cmd_line, "machinefile", 0, 0);
+            value = opal_cmd_line_get_param(&cmd_line, "machinefile", 0, 0);
             orte_set_attribute(&app->attributes, ORTE_APP_HOSTFILE, ORTE_ATTR_GLOBAL, value, OPAL_STRING);
         }
     }
 
     /* Did the user specify any hosts? */
-    if (0 < (j = opal_cmd_line_get_ninsts(&app_cmd_line, "host"))) {
+    if (0 < (j = opal_cmd_line_get_ninsts(&cmd_line, "host"))) {
         char **targ=NULL, *tval;
         for (i = 0; i < j; ++i) {
-            value = opal_cmd_line_get_param(&app_cmd_line, "host", i, 0);
+            value = opal_cmd_line_get_param(&cmd_line, "host", i, 0);
             opal_argv_append_nosize(&targ, value);
         }
         tval = opal_argv_join(targ, ',');
@@ -1566,7 +1495,7 @@ static int create_app(int argc, char* argv[],
         OBJ_RELEASE(app);
     }
     if (cmd_line_made) {
-        OBJ_DESTRUCT(&app_cmd_line);
+        OBJ_DESTRUCT(&cmd_line);
     }
     return rc;
 }
@@ -1725,6 +1654,20 @@ static int parse_appfile(orte_job_t *jdata, char *filename, char ***env)
     return ORTE_SUCCESS;
 }
 
+void orte_timeout_wakeup(int sd, short args, void *cbdata)
+{
+    char *tm;
+
+    /* this function gets called when the job execution time
+     * has hit a prescribed limit - so just abort
+     */
+    tm = getenv("MPIEXEC_TIMEOUT");
+    orte_show_help("help-orterun.txt", "orterun:timeout",
+                   true, (NULL == tm) ? "NULL" : tm);
+    ORTE_UPDATE_EXIT_STATUS(ORTE_ERROR_DEFAULT_EXIT_CODE);
+    exit(orte_exit_status);
+}
+
 static void launch_recv(int status, orte_process_name_t* sender,
                         opal_buffer_t *buffer,
                         orte_rml_tag_t tag, void *cbdata)
@@ -1864,881 +1807,4 @@ static void complete_recv(int status, orte_process_name_t* sender,
     /* cleanup */
     opal_pointer_array_set_item(&tool_jobs, tool_job_index, NULL);
     OBJ_RELEASE(trk);
-}
-
-
-/****    DEBUGGER CODE ****/
-/*
- * Debugger support for orterun
- *
- * We interpret the MPICH debugger interface as follows:
- *
- * a) The launcher
- *      - spawns the other processes,
- *      - fills in the table MPIR_proctable, and sets MPIR_proctable_size
- *      - sets MPIR_debug_state to MPIR_DEBUG_SPAWNED ( = 1)
- *      - calls MPIR_Breakpoint() which the debugger will have a
- *    breakpoint on.
- *
- *  b) Applications start and then spin until MPIR_debug_gate is set
- *     non-zero by the debugger.
- *
- * This file implements (a).
- *
- **************************************************************************
- *
- * Note that we have presently tested both TotalView and DDT parallel
- * debuggers.  They both nominally subscribe to the Etnus attaching
- * interface, but there are differences between the two.
- *
- * TotalView: user launches "totalview mpirun -a ...<mpirun args>...".
- * TV launches mpirun.  mpirun launches the application and then calls
- * MPIR_Breakpoint().  This is the signal to TV that it's a parallel
- * MPI job.  TV then reads the proctable in mpirun and attaches itself
- * to all the processes (it takes care of launching itself on the
- * remote nodes).  Upon attaching to all the MPI processes, the
- * variable MPIR_being_debugged is set to 1.  When it has finished
- * attaching itself to all the MPI processes that it wants to,
- * MPIR_Breakpoint() returns.
- *
- * DDT: user launches "ddt bin -np X <mpi app name>".  DDT fork/exec's
- * mpirun to launch ddt-debugger on the back-end nodes via "mpirun -np
- * X ddt-debugger" (not the lack of other arguments -- we can't pass
- * anything to mpirun).  This app will eventually fork/exec the MPI
- * app.  DDT does not current set MPIR_being_debugged in the MPI app.
- *
- **************************************************************************
- *
- * We support two ways of waiting for attaching debuggers.  The
- * implementation spans this file and ompi/debuggers/ompi_debuggers.c.
- *
- * 1. If using orterun: MPI processes will have the
- * orte_in_parallel_debugger MCA param set to true (because not all
- * debuggers consistently set MPIR_being_debugged in both the launcher
- * and in the MPI procs).  The HNP will call MPIR_Breakpoint() and
- * then RML send a message to VPID 0 (MCW rank 0) when it returns
- * (MPIR_Breakpoint() doesn't return until the debugger has attached
- * to all relevant processes).  Meanwhile, VPID 0 blocks waiting for
- * the RML message.  All other VPIDs immediately call the grpcomm
- * barrier (and therefore block until the debugger attaches).  Once
- * VPID 0 receives the RML message, we know that the debugger has
- * attached to all processes that it cares about, and VPID 0 then
- * joins the grpcomm barrier, allowing the job to continue.  This
- * scheme has the side effect of nicely supporting partial attaches by
- * parallel debuggers (i.e., attaching to only some of the MPI
- * processes; not necessarily all of them).
- *
- * 2. If not using orterun: in this case, we know that there will not be an RML message
- * sent to VPID 0.  So we have to look for a magic environment
- * variable from the launcher to know if the jobs will be attached by
- * a debugger (e.g., set by yod, srun, ...etc.), and if so, spin on
- * MPIR_debug_gate.  These environment variable names must be
- * hard-coded in the OMPI layer (see ompi/debuggers/ompi_debuggers.c).
- */
-
-/* local globals and functions */
-#define DUMP_INT(X) fprintf(stderr, "  %s = %d\n", # X, X);
-#define FILE_MODE (S_IRUSR | S_IWUSR | S_IRGRP | S_IROTH)
-
-struct MPIR_PROCDESC {
-    char *host_name;        /* something that can be passed to inet_addr */
-    char *executable_name;  /* name of binary */
-    int pid;                /* process pid */
-};
-
-
-/**
- * Initialization of data structures for running under a debugger
- * using the MPICH/TotalView parallel debugger interface.  Before the
- * spawn we need to check if we are being run under a TotalView-like
- * debugger; if so then inform applications via an MCA parameter.
- */
-static void orte_debugger_init_before_spawn(orte_job_t *jdata)
-{
-    char *env_name;
-    orte_app_context_t *app;
-    int i;
-    char *attach_fifo;
-
-    if (!MPIR_being_debugged && !orte_in_parallel_debugger) {
-        /* if we were given a test debugger, then we still want to
-         * colaunch it
-         */
-        if (NULL != orte_debugger_test_daemon) {
-            opal_output_verbose(2, orte_debug_output,
-                                "%s No debugger test daemon specified",
-                                ORTE_NAME_PRINT(ORTE_PROC_MY_NAME));
-            goto launchit;
-        }
-        /* if we were given an auto-detect rate, then we want to setup
-         * an event so we periodically do the check
-         */
-        if (0 < orte_debugger_check_rate) {
-            opal_output_verbose(2, orte_debug_output,
-                                "%s Setting debugger attach check rate for %d seconds",
-                                ORTE_NAME_PRINT(ORTE_PROC_MY_NAME),
-                                orte_debugger_check_rate);
-            ORTE_TIMER_EVENT(orte_debugger_check_rate, 0, attach_debugger, ORTE_SYS_PRI);
-        } else if (orte_create_session_dirs) {
-            /* create the attachment FIFO and setup readevent - cannot be
-             * done if no session dirs exist!
-             */
-            attach_fifo = opal_os_path(false, orte_process_info.job_session_dir, "debugger_attach_fifo", NULL);
-            if ((mkfifo(attach_fifo, FILE_MODE) < 0) && errno != EEXIST) {
-                opal_output(0, "CANNOT CREATE FIFO %s: errno %d", attach_fifo, errno);
-                free(attach_fifo);
-                return;
-            }
-            strncpy(MPIR_attach_fifo, attach_fifo, MPIR_MAX_PATH_LENGTH - 1);
-            free(attach_fifo);
-            open_fifo();
-        }
-        return;
-    }
-
- launchit:
-    opal_output_verbose(1, orte_debug_output, "Info: Spawned by a debugger");
-
-    /* tell the procs they are being debugged */
-    (void) mca_base_var_env_name ("orte_in_parallel_debugger", &env_name);
-
-    for (i=0; i < jdata->apps->size; i++) {
-        if (NULL == (app = (orte_app_context_t*)opal_pointer_array_get_item(jdata->apps, i))) {
-            continue;
-        }
-        opal_setenv(env_name, "1", true, &app->env);
-    }
-    free(env_name);
-}
-
-static bool mpir_breakpoint_fired = false;
-
-static void orte_debugger_dump(void)
-{
-    int i;
-
-    DUMP_INT(MPIR_being_debugged);
-    DUMP_INT(MPIR_debug_state);
-    DUMP_INT(MPIR_partial_attach_ok);
-    DUMP_INT(MPIR_i_am_starter);
-    DUMP_INT(MPIR_forward_output);
-    DUMP_INT(MPIR_proctable_size);
-    fprintf(stderr, "  MPIR_proctable:\n");
-    for (i = 0; i < MPIR_proctable_size; i++) {
-        fprintf(stderr,
-                "    (i, host, exe, pid) = (%d, %s, %s, %d)\n",
-                i,
-                MPIR_proctable[i].host_name,
-                MPIR_proctable[i].executable_name,
-                MPIR_proctable[i].pid);
-    }
-    fprintf(stderr, "MPIR_executable_path: %s\n",
-            ('\0' == MPIR_executable_path[0]) ?
-            "NULL" : (char*) MPIR_executable_path);
-    fprintf(stderr, "MPIR_server_arguments: %s\n",
-            ('\0' == MPIR_server_arguments[0]) ?
-            "NULL" : (char*) MPIR_server_arguments);
-}
-
-static void setup_debugger_job(void)
-{
-    orte_job_t *debugger;
-    orte_app_context_t *app;
-    orte_proc_t *proc;
-    int i, rc;
-    orte_node_t *node;
-    orte_vpid_t vpid=0;
-    char cwd[OPAL_PATH_MAX];
-
-    /* setup debugger daemon job */
-    debugger = OBJ_NEW(orte_job_t);
-    /* create a jobid for these daemons - this is done solely
-     * to avoid confusing the rest of the system's bookkeeping
-     */
-    orte_plm_base_create_jobid(debugger);
-    /* set the personality to ORTE */
-    opal_argv_append_nosize(&debugger->personality, "orte");
-    /* flag the job as being debugger daemons */
-    ORTE_FLAG_SET(debugger, ORTE_JOB_FLAG_DEBUGGER_DAEMON);
-    /* unless directed, we do not forward output */
-    if (!MPIR_forward_output) {
-        ORTE_FLAG_SET(debugger, ORTE_JOB_FLAG_FORWARD_OUTPUT);
-    }
-    /* dont push stdin */
-    debugger->stdin_target = ORTE_VPID_INVALID;
-    /* add it to the global job pool */
-    opal_hash_table_set_value_uint32(orte_job_data, debugger->jobid, debugger);
-    /* create an app_context for the debugger daemon */
-    app = OBJ_NEW(orte_app_context_t);
-    if (NULL != orte_debugger_test_daemon) {
-        app->app = strdup(orte_debugger_test_daemon);
-    } else {
-        app->app = strdup((char*)MPIR_executable_path);
-    }
-    /* don't currently have an option to pass the debugger
-     * cwd - probably should add one someday
-     */
-    if (OPAL_SUCCESS != (rc = opal_getcwd(cwd, sizeof(cwd)))) {
-        orte_show_help("help-orterun.txt", "orterun:init-failure",
-                       true, "get the cwd", rc);
-        return;
-    }
-    app->cwd = strdup(cwd);
-    orte_remove_attribute(&app->attributes, ORTE_APP_USER_CWD);
-    opal_argv_append_nosize(&app->argv, app->app);
-    build_debugger_args(app);
-    opal_pointer_array_add(debugger->apps, app);
-    debugger->num_apps = 1;
-    /* create a job map */
-    debugger->map = OBJ_NEW(orte_job_map_t);
-    /* in building the map, we want to launch one debugger daemon
-     * on each node that *already has an application process on it*.
-     * We cannot just launch one debugger daemon on EVERY node because
-     * the original job may not have placed procs on every node. So
-     * we construct the map here by cycling across all nodes, adding
-     * only those nodes where num_procs > 0.
-     */
-    for (i=0; i < orte_node_pool->size; i++) {
-        if (NULL == (node = (orte_node_t*)opal_pointer_array_get_item(orte_node_pool, i))) {
-            continue;
-        }
-        /* if this node wasn't included in the vm, ignore it */
-        if (NULL == node->daemon) {
-            continue;
-        }
-        /* if the node doesn't have any app procs on it, ignore it */
-        if (node->num_procs < 1) {
-            continue;
-        }
-        /* this node has at least one proc, so add it to our map */
-        OBJ_RETAIN(node);
-        opal_pointer_array_add(debugger->map->nodes, node);
-        debugger->map->num_nodes++;
-        /* add a debugger daemon to the node - note that the
-         * debugger daemon does NOT count against our subscribed slots
-         */
-        proc = OBJ_NEW(orte_proc_t);
-        proc->name.jobid = debugger->jobid;
-        proc->name.vpid = vpid++;
-        /* set the local/node ranks - we don't actually care
-         * what these are, but the odls needs them
-         */
-        proc->local_rank = 0;
-        proc->node_rank = 0;
-        proc->app_rank = proc->name.vpid;
-        /* flag the proc as ready for launch */
-        proc->state = ORTE_PROC_STATE_INIT;
-        proc->app_idx = 0;
-
-        OBJ_RETAIN(node);  /* maintain accounting on object */
-        proc->node = node;
-        /* add the proc to the job */
-        opal_pointer_array_set_item(debugger->procs, proc->name.vpid, proc);
-        debugger->num_procs++;
-
-        /* add the proc to the node's array */
-        OBJ_RETAIN(proc);
-        opal_pointer_array_add(node->procs, (void*)proc);
-        node->num_procs++;
-    }
-    /* schedule it for launch */
-    debugger->state = ORTE_JOB_STATE_INIT;
-    ORTE_ACTIVATE_JOB_STATE(debugger, ORTE_JOB_STATE_LAUNCH_APPS);
-}
-
-/*
- * Initialization of data structures for running under a debugger
- * using the MPICH/TotalView parallel debugger interface. This stage
- * of initialization must occur after spawn
- *
- * NOTE: We -always- perform this step to ensure that any debugger
- * that attaches to us post-launch of the application can get a
- * completed proctable
- */
-void orte_debugger_init_after_spawn(int fd, short event, void *cbdata)
-{
-    orte_state_caddy_t *caddy = (orte_state_caddy_t*)cbdata;
-    orte_job_t *jdata = caddy->jdata;
-    orte_proc_t *proc;
-    orte_app_context_t *appctx;
-    orte_vpid_t i, j;
-    opal_buffer_t *buf;
-    int rc;
-    char **aliases, *aptr;
-
-    /* if we couldn't get thru the mapper stage, we might
-     * enter here with no procs. Avoid the "zero byte malloc"
-     * message by checking here
-     */
-    if (MPIR_proctable || 0 == jdata->num_procs) {
-        /* already initialized */
-        opal_output_verbose(5, orte_debug_output,
-                            "%s: debugger already initialized or zero procs",
-                            ORTE_NAME_PRINT(ORTE_PROC_MY_NAME));
-        OBJ_RELEASE(caddy);
-        if (!mpir_breakpoint_fired) {
-            /* record that we have triggered the debugger */
-            mpir_breakpoint_fired = true;
-
-            /* trigger the debugger */
-            MPIR_Breakpoint();
-
-            /* send a message to rank=0 to release it */
-            if (NULL == (proc = (orte_proc_t*)opal_pointer_array_get_item(jdata->procs, 0)) ||
-                ORTE_PROC_STATE_UNTERMINATED < proc->state ) {
-                /* proc is already dead */
-                return;
-            }
-            buf = OBJ_NEW(opal_buffer_t); /* don't need anything in this */
-            if (0 > (rc = orte_rml.send_buffer_nb(&proc->name, buf,
-                                                  ORTE_RML_TAG_DEBUGGER_RELEASE,
-                                                  orte_rml_send_callback, NULL))) {
-                opal_output(0, "Error: could not send debugger release to MPI procs - error %s", ORTE_ERROR_NAME(rc));
-                OBJ_RELEASE(buf);
-            }
-        }
-        return;
-    }
-
-    /* fill in the proc table for the application processes */
-
-    opal_output_verbose(5, orte_debug_output,
-                        "%s: Setting up debugger process table for applications",
-                        ORTE_NAME_PRINT(ORTE_PROC_MY_NAME));
-
-    MPIR_debug_state = 1;
-
-    /* set the total number of processes in the job */
-    MPIR_proctable_size = jdata->num_procs;
-
-    /* allocate MPIR_proctable */
-    MPIR_proctable = (struct MPIR_PROCDESC *)malloc(sizeof(struct MPIR_PROCDESC) *
-                                                    MPIR_proctable_size);
-    if (MPIR_proctable == NULL) {
-        ORTE_ERROR_LOG(ORTE_ERR_OUT_OF_RESOURCE);
-        OBJ_RELEASE(caddy);
-        return;
-    }
-
-    if (orte_debugger_dump_proctable) {
-        opal_output(orte_clean_output, "MPIR Proctable for job %s", ORTE_JOBID_PRINT(jdata->jobid));
-    }
-
-    /* initialize MPIR_proctable */
-    for (j=0; j < jdata->num_procs; j++) {
-        if (NULL == (proc = (orte_proc_t*)opal_pointer_array_get_item(jdata->procs, j))) {
-            continue;
-        }
-        /* store this data in the location whose index
-         * corresponds to the proc's rank
-         */
-        i = proc->name.vpid;
-        if (NULL == (appctx = (orte_app_context_t*)opal_pointer_array_get_item(jdata->apps, proc->app_idx))) {
-            continue;
-        }
-
-        /* take the indicated alias as the hostname, if aliases exist */
-        if (orte_retain_aliases) {
-            aliases = NULL;
-            aptr = NULL;
-            if (orte_get_attribute(&proc->node->attributes, ORTE_NODE_ALIAS, (void**)&aptr, OPAL_STRING)) {
-                aliases = opal_argv_split(aptr, ',');
-                free(aptr);
-                if (orte_use_hostname_alias <= opal_argv_count(aliases)) {
-                    MPIR_proctable[i].host_name = strdup(aliases[orte_use_hostname_alias-1]);
-                }
-                opal_argv_free(aliases);
-            }
-        } else {
-            /* just use the default name */
-            MPIR_proctable[i].host_name = strdup(proc->node->name);
-        }
-
-        if ( 0 == strncmp(appctx->app, OPAL_PATH_SEP, 1 )) {
-            MPIR_proctable[i].executable_name =
-                opal_os_path( false, appctx->app, NULL );
-        } else {
-            MPIR_proctable[i].executable_name =
-                opal_os_path( false, appctx->cwd, appctx->app, NULL );
-        }
-        MPIR_proctable[i].pid = proc->pid;
-        if (orte_debugger_dump_proctable) {
-            opal_output(orte_clean_output, "%s: Host %s Exe %s Pid %d",
-                        ORTE_VPID_PRINT(i), MPIR_proctable[i].host_name,
-                        MPIR_proctable[i].executable_name, MPIR_proctable[i].pid);
-        }
-    }
-
-    if (0 < opal_output_get_verbosity(orte_debug_output)) {
-        orte_debugger_dump();
-    }
-
-    /* if we are being launched under a debugger, then we must wait
-     * for it to be ready to go and do some things to start the job
-     */
-    if (MPIR_being_debugged || NULL != orte_debugger_test_daemon ||
-        NULL != getenv("ORTE_TEST_DEBUGGER_ATTACH")) {
-        /* if we are not launching debugger daemons, then trigger
-         * the debugger - otherwise, we need to wait for the debugger
-         * daemons to be started
-         */
-        if ('\0' == MPIR_executable_path[0] && NULL == orte_debugger_test_daemon) {
-            /* record that we have triggered the debugger */
-            mpir_breakpoint_fired = true;
-
-            /* trigger the debugger */
-            MPIR_Breakpoint();
-
-            /* send a message to rank=0 to release it */
-            if (NULL == (proc = (orte_proc_t*)opal_pointer_array_get_item(jdata->procs, 0)) ||
-                ORTE_PROC_STATE_UNTERMINATED < proc->state) {
-                /* proc is already dead or never registered with us (so we don't have
-                 * contact info for him)
-                 */
-                return;
-            }
-            opal_output_verbose(2, orte_debug_output,
-                                "%s sending debugger release to %s",
-                                ORTE_NAME_PRINT(ORTE_PROC_MY_NAME),
-                                ORTE_NAME_PRINT(&proc->name));
-            buf = OBJ_NEW(opal_buffer_t); /* don't need anything in this */
-            if (0 > (rc = orte_rml.send_buffer_nb(&proc->name, buf,
-                                                  ORTE_RML_TAG_DEBUGGER_RELEASE,
-                                                  orte_rml_send_callback, NULL))) {
-                opal_output(0, "Error: could not send debugger release to MPI procs - error %s", ORTE_ERROR_NAME(rc));
-                OBJ_RELEASE(buf);
-            }
-        } else {
-            /* if I am launching debugger daemons, then I need to do so now
-             * that the job has been started and I know which nodes have
-             * apps on them
-             */
-            opal_output_verbose(2, orte_debug_output,
-                                "%s Cospawning debugger daemons %s",
-                                ORTE_NAME_PRINT(ORTE_PROC_MY_NAME),
-                                (NULL == orte_debugger_test_daemon) ?
-                                MPIR_executable_path : orte_debugger_test_daemon);
-            setup_debugger_job();
-        }
-        /* we don't have anything else to do */
-        OBJ_RELEASE(caddy);
-        return;
-    }
-
-    /* if we are not being debugged, then just cleanup and depart */
-    OBJ_RELEASE(caddy);
-}
-
-/*
- * Process one line from the orte_base_user_debugger MCA param and
- * look for that debugger in the path.  If we find it, fill in
- * new_argv.
- */
-static int process(char *orig_line, char *basename, opal_cmd_line_t *cmd_line,
-                   int argc, char **argv, char ***new_argv, int num_procs)
-{
-    int ret = ORTE_SUCCESS;
-    int i, j, count;
-    char *line = NULL, *tmp = NULL, *full_line = strdup(orig_line);
-    char **orterun_argv = NULL, **executable_argv = NULL, **line_argv = NULL;
-    char cwd[OPAL_PATH_MAX];
-    bool used_num_procs = false;
-    bool single_app = false;
-    bool fail_needed_executable = false;
-
-    line = full_line;
-    if (NULL == line) {
-        ret = ORTE_ERR_OUT_OF_RESOURCE;
-        goto out;
-    }
-
-    /* Trim off whitespace at the beginning and ending of line */
-
-    for (i = 0; '\0' != line[i] && isspace(line[i]); ++line) {
-        continue;
-    }
-    for (i = strlen(line) - 2; i > 0 && isspace(line[i]); ++i) {
-        line[i] = '\0';
-    }
-    if (strlen(line) <= 0) {
-        ret = ORTE_ERROR;
-        goto out;
-    }
-
-    /* Get the tail of the command line (i.e., the user executable /
-       argv) */
-
-    opal_cmd_line_get_tail(cmd_line, &i, &executable_argv);
-
-    /* Make a new copy of the orterun command line args, without the
-       orterun token itself, and without the --debug, --debugger, and
-       -tv flags. */
-
-    orterun_argv = opal_argv_copy(argv);
-    count = opal_argv_count(orterun_argv);
-    opal_argv_delete(&count, &orterun_argv, 0, 1);
-    for (i = 0; NULL != orterun_argv[i]; ++i) {
-        count = opal_argv_count(orterun_argv);
-        if (0 == strcmp(orterun_argv[i], "-debug") ||
-            0 == strcmp(orterun_argv[i], "--debug")) {
-            opal_argv_delete(&count, &orterun_argv, i, 1);
-        } else if (0 == strcmp(orterun_argv[i], "-tv") ||
-                   0 == strcmp(orterun_argv[i], "--tv")) {
-            opal_argv_delete(&count, &orterun_argv, i, 1);
-        } else if (0 == strcmp(orterun_argv[i], "--debugger") ||
-                   0 == strcmp(orterun_argv[i], "-debugger")) {
-            opal_argv_delete(&count, &orterun_argv, i, 2);
-        }
-    }
-
-    /* Replace @@ tokens - line should never realistically be bigger
-       than MAX_INT, so just cast to int to remove compiler warning */
-
-    *new_argv = NULL;
-    line_argv = opal_argv_split(line, ' ');
-    if (NULL == line_argv) {
-        ret = ORTE_ERR_NOT_FOUND;
-        goto out;
-    }
-    for (i = 0; NULL != line_argv[i]; ++i) {
-        if (0 == strcmp(line_argv[i], "@mpirun@") ||
-            0 == strcmp(line_argv[i], "@orterun@")) {
-            opal_argv_append_nosize(new_argv, argv[0]);
-        } else if (0 == strcmp(line_argv[i], "@mpirun_args@") ||
-                   0 == strcmp(line_argv[i], "@orterun_args@")) {
-            for (j = 0; NULL != orterun_argv && NULL != orterun_argv[j]; ++j) {
-                opal_argv_append_nosize(new_argv, orterun_argv[j]);
-            }
-        } else if (0 == strcmp(line_argv[i], "@np@")) {
-            used_num_procs = true;
-            asprintf(&tmp, "%d", num_procs);
-            opal_argv_append_nosize(new_argv, tmp);
-            free(tmp);
-        } else if (0 == strcmp(line_argv[i], "@single_app@")) {
-            /* This token is only a flag; it is not replaced with any
-               alternate text */
-            single_app = true;
-        } else if (0 == strcmp(line_argv[i], "@executable@")) {
-            /* If we found the executable, paste it in.  Otherwise,
-               this is a possible error. */
-            if (NULL != executable_argv) {
-                opal_argv_append_nosize(new_argv, executable_argv[0]);
-            } else {
-                fail_needed_executable = true;
-            }
-        } else if (0 == strcmp(line_argv[i], "@executable_argv@")) {
-            /* If we found the tail, paste in the argv.  Otherwise,
-               this is a possible error. */
-            if (NULL != executable_argv) {
-                for (j = 1; NULL != executable_argv[j]; ++j) {
-                    opal_argv_append_nosize(new_argv, executable_argv[j]);
-                }
-            } else {
-                fail_needed_executable = true;
-            }
-        } else {
-            /* It wasn't a special token, so just copy it over */
-            opal_argv_append_nosize(new_argv, line_argv[i]);
-        }
-    }
-
-    /* Can we find argv[0] in the path? */
-
-    getcwd(cwd, OPAL_PATH_MAX);
-    tmp = opal_path_findv((*new_argv)[0], X_OK, environ, cwd);
-    if (NULL != tmp) {
-        free(tmp);
-
-        /* Ok, we found a good debugger.  Check for some error
-           conditions. */
-        tmp = opal_argv_join(argv, ' ');
-
-        /* We do not support launching a debugger that requires the
-           -np value if the user did not specify -np on the command
-           line. */
-        if (used_num_procs && 0 == num_procs) {
-            free(tmp);
-            tmp = opal_argv_join(orterun_argv, ' ');
-            orte_show_help("help-orterun.txt", "debugger requires -np",
-                           true, (*new_argv)[0], argv[0], tmp,
-                           (*new_argv)[0]);
-            /* Fall through to free / fail, below */
-        }
-
-        /* Some debuggers do not support launching MPMD */
-        else if (single_app && NULL != strstr(tmp, " : ")) {
-            orte_show_help("help-orterun.txt",
-                           "debugger only accepts single app", true,
-                           (*new_argv)[0], (*new_argv)[0]);
-            /* Fall through to free / fail, below */
-        }
-
-        /* Some debuggers do not use orterun/mpirun, and therefore
-           must have an executable to run (e.g., cannot use mpirun's
-           app context file feature). */
-        else if (fail_needed_executable) {
-            orte_show_help("help-orterun.txt",
-                           "debugger requires executable", true,
-                           (*new_argv)[0], argv[0], (*new_argv)[0], argv[0],
-                           (*new_argv)[0]);
-            /* Fall through to free / fail, below */
-        }
-
-        /* Otherwise, we succeeded.  Return happiness. */
-        else {
-            goto out;
-        }
-    }
-
-    /* All done -- didn't find it */
-
-    opal_argv_free(*new_argv);
-    *new_argv = NULL;
-    ret = ORTE_ERR_NOT_FOUND;
-
- out:
-    if (NULL != orterun_argv) {
-        opal_argv_free(orterun_argv);
-    }
-    if (NULL != executable_argv) {
-        opal_argv_free(executable_argv);
-    }
-    if (NULL != line_argv) {
-        opal_argv_free(line_argv);
-    }
-    if (NULL != tmp) {
-        free(tmp);
-    }
-    if (NULL != full_line) {
-        free(full_line);
-    }
-    return ret;
-}
-
-static void open_fifo (void)
-{
-    if (orte_debugger_attach_fd > 0) {
-    close(orte_debugger_attach_fd);
-    }
-
-    orte_debugger_attach_fd = open(MPIR_attach_fifo, O_RDONLY | O_NONBLOCK, 0);
-    if (orte_debugger_attach_fd < 0) {
-    opal_output(0, "%s unable to open debugger attach fifo",
-            ORTE_NAME_PRINT(ORTE_PROC_MY_NAME));
-    return;
-    }
-
-    /* Set this fd to be close-on-exec so that children don't see it */
-    if (opal_fd_set_cloexec(orte_debugger_attach_fd) != OPAL_SUCCESS) {
-        opal_output(0, "%s unable to set debugger attach fifo to CLOEXEC",
-                    ORTE_NAME_PRINT(ORTE_PROC_MY_NAME));
-        close(orte_debugger_attach_fd);
-        orte_debugger_attach_fd = -1;
-        return;
-    }
-
-    opal_output_verbose(2, orte_debug_output,
-            "%s Monitoring debugger attach fifo %s",
-            ORTE_NAME_PRINT(ORTE_PROC_MY_NAME),
-            MPIR_attach_fifo);
-    orte_debugger_attach = (opal_event_t*)malloc(sizeof(opal_event_t));
-    opal_event_set(orte_event_base, orte_debugger_attach, orte_debugger_attach_fd,
-                   OPAL_EV_READ, attach_debugger, orte_debugger_attach);
-
-    orte_debugger_fifo_active = true;
-    opal_event_add(orte_debugger_attach, 0);
-}
-
-static void attach_debugger(int fd, short event, void *arg)
-{
-    unsigned char fifo_cmd;
-    int rc;
-    orte_timer_t *tm;
-
-    if (orte_debugger_fifo_active) {
-        orte_debugger_attach = (opal_event_t*)arg;
-        orte_debugger_fifo_active = false;
-
-        rc = read(orte_debugger_attach_fd, &fifo_cmd, sizeof(fifo_cmd));
-        if (!rc) {
-            /* release the current event */
-            opal_event_free(orte_debugger_attach);
-            /* reopen device to clear hangup */
-            open_fifo();
-            return;
-        }
-        if (1 != fifo_cmd) {
-            /* ignore the cmd */
-            orte_debugger_fifo_active = true;
-            opal_event_add(orte_debugger_attach, 0);
-            return;
-        }
-    }
-
-    if (!MPIR_being_debugged && !orte_debugger_test_attach) {
-        /* false alarm - reset the read or timer event */
-        if (0 == orte_debugger_check_rate) {
-            orte_debugger_fifo_active = true;
-            opal_event_add(orte_debugger_attach, 0);
-        } else if (!MPIR_being_debugged) {
-            tm = (orte_timer_t*)arg;
-            /* re-add the event */
-            opal_event_evtimer_add(tm->ev, &tm->tv);
-        }
-        return;
-    }
-
-    opal_output_verbose(1, orte_debug_output,
-                        "%s Attaching debugger %s", ORTE_NAME_PRINT(ORTE_PROC_MY_NAME),
-                        (NULL == orte_debugger_test_daemon) ? MPIR_executable_path : orte_debugger_test_daemon);
-
-    /* a debugger has attached! All the MPIR_Proctable
-     * data is already available, so we only need to
-     * check to see if we should spawn any daemons
-     */
-    if ('\0' != MPIR_executable_path[0] || NULL != orte_debugger_test_daemon) {
-        opal_output_verbose(2, orte_debug_output,
-                            "%s Spawning debugger daemons %s",
-                            ORTE_NAME_PRINT(ORTE_PROC_MY_NAME),
-                            (NULL == orte_debugger_test_daemon) ?
-                            MPIR_executable_path : orte_debugger_test_daemon);
-        setup_debugger_job();
-    }
-
-    /* reset the read or timer event */
-    if (0 == orte_debugger_check_rate) {
-        orte_debugger_fifo_active = true;
-        opal_event_add(orte_debugger_attach, 0);
-    } else if (!MPIR_being_debugged) {
-        tm = (orte_timer_t*)arg;
-        /* re-add the event */
-        opal_event_evtimer_add(tm->ev, &tm->tv);
-    }
-}
-
-static void build_debugger_args(orte_app_context_t *debugger)
-{
-    int i, j;
-    char mpir_arg[MPIR_MAX_ARG_LENGTH];
-
-    if ('\0' != MPIR_server_arguments[0]) {
-        j=0;
-        memset(mpir_arg, 0, MPIR_MAX_ARG_LENGTH);
-        for (i=0; i < MPIR_MAX_ARG_LENGTH; i++) {
-            if (MPIR_server_arguments[i] == '\0') {
-                if (0 < j) {
-                    opal_argv_append_nosize(&debugger->argv, mpir_arg);
-                    memset(mpir_arg, 0, MPIR_MAX_ARG_LENGTH);
-                    j=0;
-                }
-            } else {
-                mpir_arg[j] = MPIR_server_arguments[i];
-                j++;
-            }
-        }
-    }
-}
-
-/**
- * Run a user-level debugger
- */
-static void run_debugger(char *basename, opal_cmd_line_t *cmd_line,
-                         int argc, char *argv[], int num_procs)
-{
-    int i, id, ret;
-    char **new_argv = NULL;
-    const char **tmp;
-    char *value, **lines, *env_name;
-
-    /* Get the orte_base_debug MCA parameter and search for a debugger
-       that can run */
-
-    id = mca_base_var_find("orte", "orte", NULL, "base_user_debugger");
-    if (id < 0) {
-        orte_show_help("help-orterun.txt", "debugger-mca-param-not-found",
-                       true);
-        exit(1);
-    }
-
-    ret = mca_base_var_get_value (id, &tmp, NULL, NULL);
-    if (OPAL_SUCCESS != ret || NULL == tmp || NULL == tmp[0]) {
-        orte_show_help("help-orterun.txt", "debugger-orte_base_user_debugger-empty",
-                       true);
-        exit(1);
-    }
-
-    /* Look through all the values in the MCA param */
-
-    lines = opal_argv_split(tmp[0], ':');
-    for (i = 0; NULL != lines[i]; ++i) {
-        if (ORTE_SUCCESS == process(lines[i], basename, cmd_line, argc, argv,
-                                    &new_argv, num_procs)) {
-            break;
-        }
-    }
-
-    /* If we didn't find one, abort */
-
-    if (NULL == lines[i]) {
-        orte_show_help("help-orterun.txt", "debugger-not-found", true);
-        exit(1);
-    }
-    opal_argv_free(lines);
-
-    /* We found one */
-
-    /* cleanup the MPIR arrays in case the debugger doesn't set them */
-    memset((char*)MPIR_executable_path, 0, MPIR_MAX_PATH_LENGTH);
-    memset((char*)MPIR_server_arguments, 0, MPIR_MAX_ARG_LENGTH);
-
-    /* Set an MCA param so that everyone knows that they are being
-       launched under a debugger; not all debuggers are consistent
-       about setting MPIR_being_debugged in both the launcher and the
-       MPI processes */
-    ret = mca_base_var_env_name ("orte_in_parallel_debugger", &env_name);
-    if (OPAL_SUCCESS == ret && NULL != env_name) {
-        opal_setenv(env_name, "1", true, &environ);
-        free(env_name);
-    }
-
-    /* Launch the debugger */
-    execvp(new_argv[0], new_argv);
-    value = opal_argv_join(new_argv, ' ');
-    orte_show_help("help-orterun.txt", "debugger-exec-failed",
-                   true, basename, value, new_argv[0]);
-    free(value);
-    opal_argv_free(new_argv);
-    exit(1);
-}
-
-void orte_debugger_detached(int fd, short event, void *cbdata)
-{
-    orte_state_caddy_t *caddy = (orte_state_caddy_t*)cbdata;
-    OBJ_RELEASE(caddy);
-
-    /* need to ensure MPIR_Breakpoint is called again if another debugger attaches */
-    mpir_breakpoint_fired = false;
-}
-
-void orte_timeout_wakeup(int sd, short args, void *cbdata)
-{
-    char *tm;
-
-    /* this function gets called when the job execution time
-     * has hit a prescribed limit - so just abort
-     */
-    tm = getenv("MPIEXEC_TIMEOUT");
-    orte_show_help("help-orterun.txt", "orterun:timeout",
-                   true, (NULL == tm) ? "NULL" : tm);
-    ORTE_UPDATE_EXIT_STATUS(ORTE_ERROR_DEFAULT_EXIT_CODE);
-    /* if we are testing HNP suicide, then just exit */
-    if (ORTE_PROC_IS_HNP &&
-        NULL != getenv("ORTE_TEST_HNP_SUICIDE")) {
-        opal_output(0, "HNP exiting w/o cleanup");
-        exit(1);
-    }
-    /* abort the job */
-    ORTE_ACTIVATE_JOB_STATE(NULL, ORTE_JOB_STATE_ALL_JOBS_COMPLETE);
-    /* set the global abnormal exit flag  */
-    orte_abnormal_term_ordered = true;
 }
