+++ orte/tools/Makefile.am
@@ -13,7 +13,7 @@
 # Copyright (c) 2006-2008 Cisco Systems, Inc.  All rights reserved.
 # Copyright (c) 2011-2013 Los Alamos National Security, LLC.  All rights
 #                         reserved.
-# Copyright (c) 2014-2016 Intel, Inc.  All rights reserved.
+# Copyright (c) 2014-2015 Intel, Inc.  All rights reserved.
 # $COPYRIGHT$
 #
 # Additional copyrights may follow
@@ -36,6 +36,7 @@ SUBDIRS += \
         tools/orte-info \
         tools/orte-migrate \
         tools/orte-server \
+        tools/orte-submit \
         tools/orte-dvm
 
 DIST_SUBDIRS += \
@@ -50,5 +51,6 @@ DIST_SUBDIRS += \
         tools/orte-info \
         tools/orte-migrate \
         tools/orte-server \
+        tools/orte-submit \
         tools/orte-dvm
